### 2015-04-08 #10 @Robinlovelace
stplanr
* 1. What does this package do? (explain in 50 words or less)
This package brings together a number of tools that enable transport modelling and planning and R, with an emphasis on analyses needed for sustainability (e.g. planning new bicycle paths).
* 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).

```
Package: stplanr
Title: Sustainable Transport Planning with R
Version: 0.0.1.1
    person("Robin", "Lovelace", email = "rob00x@gmail.com", role = c("aut", "cre")),
    person("Richard", "Ellison", role = c("aut"), comment = "Author of various functions"),
    person("Barry", "Rowlingson", role = c("aut"), comment = "Author of overline"),
    person("Nick", "Bearman", role = c("aut"), comment = "Co-author of gclip")
    )
Description: Functionality and data access tools for transport planning, including origin-destination analysis, route allocation and modelling travel patterns.
LazyData: yes
    sp, R (>= 3.0)
    maptools,
    raster,
    rgdal,
    rgeos,
    RgoogleMaps,
    openxlsx,
    leaflet,
    httr
    tmap
```

* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/Robinlovelace/stplanr
* 4. What data source(s) does it work with (if applicable)?
Example data on route allocation and travel flows (included - more planned), geographical boundaries, transport networks, example travel survey data (to be added subsequently)
* 5. Who is the target audience?
Transport researchers and modellers from academic, public and private sectors. University students studying transport geography, transport modelling and related disciplines.
* 6. Check the box next to each policy below, confirming that you agree. These are mandatory.
* [x] The repository has continuous integration with Travis and/or another service
* [x] The package contains a vignette
* [x] The package contains a reasonably complete readme with devtools install instructions
* [x] The package contains unit tests
* 7. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
* [ ] Are there any packages that your package depends on that are not on CRAN?
* [x] Does the package have a CRAN accepted license?
* [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.

* 8. Please add explanations below for any exceptions to the above:
This is work in progress and I am learning how to build R packages as I progress. These issues will be fixed, probably during a course on R package development I will attend.
* 9. If this is a resubmission following rejection, please explain the change in cirucmstances.
