### 2015-04-13 bc9910c
Updates for GIS4TA workshop
