+++ DESCRIPTION
@@ -6,6 +6,7 @@ Date: 2015-03-10
 Author: Robin Lovelace
 Maintainer: <rob00x@gmail.com>
 Description: This is an R package for sustainable transport planning in R (stplanr).
+Depends: R (>= 3.1)
 License: MIT
 Suggests:
     ggmap,
