### 2015-04-13 #10 @sckott
hi @Robinlovelace - Thanks for your submission.

Did you mean to check off some of those things?  E.g., you checked that you have continuous integration, but you don't, and that you have tests, but at least we don't see any.

Can you correct the information above?
