### 2015-04-13 #10 @sckott
@Robinlovelace Up to you if you want to close the issue and open a new one later.

Where are users expected to get data for use in functions from your package? If data is publicly available (e.g., via web APIs, or even just ftp), we can help build in some functions to provide data access.

> I'm attending a 3 day course on R package development in May.

cool, sounds good!

The `R CMD CHECK` warnings aren't a huge deal - we can help with those. Bigger issues are:

- fit for ropensci - if there is data access functionality in your package (discussed above), it would be a better fit for us
- vignettes/help for users: this is pretty crucial - we want it to be easy for users to figure out how to use a package
- tests - tests are pretty crucial to make sure any functionality of the package doesn't break when something is changed
