### 2015-04-14 #10 @sckott
@Robinlovelace Okay, sounds good. So this package already does data acquisition?  Or it will? Do you mind if I uncheck boxes that aren't complete yet?
