### 2015-04-14 #10 @Robinlovelace
@sckott yes. gLines2CyclePaths harvests data from the CycleStreet.net routing API. We plan to add interfaces to other routing APIs. Here are a couple of examples of the kind of data it pulls in, of a couple of routes in Manchester and a few thousand OD flows in Coventry:
![selection_135](https://cloud.githubusercontent.com/assets/1825120/7136933/47a89522-e2ad-11e4-9887-1949447fb112.png)

![selection_179](https://cloud.githubusercontent.com/assets/1825120/7136953/7240f900-e2ad-11e4-8509-9b0f3cb248dd.png)
