### 2015-04-14 #10 @Robinlovelace
Don't mind at all if you uncheck uncompleted boxes, please do.
