### 2015-04-14 #10 @sckott
@Robinlovelace Cool - nice example!   Boxed unchecked.
