### 2015-04-15 96c77f2
Data and documention for 'cents' point dataset
