### 2015-04-15 1ccb8c3
Data and documention for 'cents' point dataset
