+++ man/flow.Rd
@@ -1,0 +1,58 @@
+% Generated by roxygen2 (4.1.0): do not edit by hand
+% Please edit documentation in R/data.r
+\docType{data}
+\name{flow}
+\alias{flow}
+\title{data frame of commuter flows}
+\format{A data frame with 49 rows and 15 columns}
+\usage{
+data(flow)
+}
+\description{
+This dataset represents commuter flows (work travel) between origin
+and destination zones (see \code{\link{cents}}).
+The data is from the UK and is available as open data:
+\url{http://wicid.ukdataservice.ac.uk/}.
+}
+\details{
+The variables are as follows:
+
+\itemize{
+  \item Area.of.residence. id of origin zone
+  \item Area.of.workplace id of destination zone
+  \item All. Travel to work flows by all modes
+  \item [,4:15]. Flows for different modes
+  \item id. unique id of flow
+}
+
+Although these variable names are unique to UK data, the data
+structure is generalisable and typical of flow data from any source.
+The key variables are the origin and destination ids, which link to
+the \code{cents} georeferenced spatial objects.
+}
+\examples{
+\dontrun{
+flow <- readRDS("~/repos/pct/pct-data/national/flow.Rds")
+data(cents)
+o <- flow$Area.of.residence \%in\% cents$geo_code[-1]
+d <- flow$Area.of.workplace \%in\% cents$geo_code[-1]
+flow <- flow[o & d, ] # subset flows with o and d in study area
+library(devtools)
+flow$id <- paste(flow$Area.of.residence, flow$Area.of.workplace)
+use_data(flow, overwrite = T)
+
+# Convert flows to SpatialLinesDataFrame
+flowlines <- gFlow2line(flow = flow, zones = cents)
+use_data(flowlines)
+
+# Convert flows to routes
+flowlines_84 <- sp::spTransform(flowlines, CRS("+init=epsg:4326"))
+routes_fast <- gLines2CyclePath(l = flowlines_84, plan = "fastest")
+routes_slow <- gLines2CyclePath(l = flowlines_84, plan = "quietest")
+
+use_data(routes_fast)
+use_data(routes_slow)
+}
+}
+\keyword{datasets}
+
