+++ man/flowlines.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2 (4.1.0): do not edit by hand
+% Please edit documentation in R/data.r
+\docType{data}
+\name{flowlines}
+\alias{flowlines}
+\title{SpatialLinesDataFrame of commuter flows}
+\format{A SpatialLinesDataFrame 49 rows and 15 columns}
+\usage{
+data(flowlines)
+}
+\description{
+Flow data after conversion to a spatial format
+with \code{\link{gFlow2line}} (see \code{\link{flow}}).
+}
+\keyword{datasets}
+
