+++ man/routes_slow.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2 (4.1.0): do not edit by hand
 % Please edit documentation in R/data.r
 \docType{data}
-\name{routes_quiet}
-\alias{routes_quiet}
+\name{routes_slow}
+\alias{routes_slow}
 \title{SpatialLinesDataFrame of commuter flows on the travel network}
 \format{A SpatialLinesDataFrame 49 rows and 15 columns}
 \usage{
@@ -15,8 +15,8 @@ objects
 with \code{\link{gFlow2line}} (see \code{\link{flow}}).
 }
 \seealso{
-\code{\link{gFlow2line}}  \code{\link{cents}}
- \code{\link{flow}}  \code{\link{routes_slow}}
+Other example.flow.data: \code{\link{cents}};
+  \code{\link{flowlines}}; \code{\link{flow}};
   \code{\link{routes_fast}}
 }
 \keyword{datasets}
