+++ man/flowlines.Rd
@@ -9,12 +9,10 @@
 data(flowlines)
 }
 \description{
-Flow data after conversion to a spatial format
-with \code{\link{gFlow2line}} (see \code{\link{flow}}).
+SpatialLinesDataFrame of commuter flows
 }
 \seealso{
-Other example.flow.data: \code{\link{cents}};
-  \code{\link{flow}}; \code{\link{routes_fast}};
+Other res: \code{\link{flow}}; \code{\link{routes_fast}};
   \code{\link{routes_slow}}
 }
 \keyword{datasets}
