### 2015-04-16 8aa1d7a
More documentation improvements
