+++ R/distance-decay-functions.R
@@ -6,7 +6,7 @@
 #'
 #'  @seealso \code{\link{dd_logsqr}}
 #' @references
-#' Iacono, M., Krizek, K. J., & El-Geneidy, A. (2010).
+#' Iacono, M., Krizek, K. J. and El-Geneidy, A. (2010).
 #' Measuring non-motorized accessibility: issues, alternatives, and execution. Journal of Transport Geography, 18(1), 133–140. doi:10.1016/j.jtrangeo.2009.02.002
 #' @examples
 #' d <- 0:10 # vector of distances
@@ -32,7 +32,7 @@ dd_iac(1:10)
 #' negative c values lead to rapid decay in cycling under 2 km.
 #'
 #' @references
-#' Iacono, M., Krizek, K. J., & El-Geneidy, A. (2010).
+#' Iacono, M., Krizek, K. J. and El-Geneidy, A. (2010).
 #' Measuring non-motorized accessibility: issues, alternatives, and execution. Journal of Transport Geography, 18(1), 133–140. doi:10.1016/j.jtrangeo.2009.02.002
 #'
 #' @examples
