+++ R/gGraphHop.R
@@ -1,3 +1,3 @@
-gGraphHop <- function(x){
-
-}
+# gGraphHop <- function(x){
+#
+# }
