### 2015-04-16 c8df924
Updated documentation - down to 3 warnings!
