+++ R/geo-functions.R
@@ -1,7 +1,14 @@
 #' Write to geojson easily
+#'
+#' Provides a user-friendly wrapper for rgdal::writeOGR().
+#'
+#' @param x The object to output
+#' @param filename File name of the output geojson
+#'
+#'
 writeGeoJSON <- function(x, filename){
   name <- nm <-deparse(substitute(x))
-  writeOGR(obj = x, layer = name, dsn = filename, driver = "GeoJSON")
+  rgdal::writeOGR(obj = x, layer = name, dsn = filename, driver = "GeoJSON")
   newname <- paste0(filename, ".geojson")
   file.rename(filename, newname)
 }
@@ -9,10 +16,15 @@ writeGeoJSON <- function(x, filename){
 #' Simplify geometry file of a shapfile.
 #'
 #' @section Details:
-#' This is a wrapper funtion for the open source JavaScript command-line GIS application mapshaper: https://github.com/mbloch/mapshaper . mapshaper which must be installed locally for gMapshaper to work. Writes \code{gMapshape} writes new file to disk. Thanks to Richard and Adrian Ellison for demonstrating this in R.
+#' This is a wrapper funtion for the open source JavaScript command-line GIS application mapshaper: \url{https://github.com/mbloch/mapshaper} . mapshaper which must be installed locally for gMapshaper to work. Writes \code{gMapshape} writes new file to disk. Thanks to Richard and Adrian Ellison for demonstrating this in R.
 #'
 #' @param dsn A character string providing the absolute path to the shapefile to simplify.
-#' @param y A number between 0 and 100 stating how aggressively to simplify the object. Roughly corresponds with the \% hard disk space saved by the simplified shapefile compared with the original.
+#' @param percent A number between 0 and 100 stating how aggressively to simplify
+#'  the object
+#' Percentage of removable points to retain.
+#' So \code{percent = 1} is a very aggressive simplication, saving a huge amount of
+#' hard-disk space.
+#'
 #' @examples
 #' \dontrun{
 #' gMapshape("~/geodata/myShapefile.shp", 5)
