+++ R/gFlow.R
@@ -16,7 +16,7 @@
 #' representing origins and destinations of travel flows.
 #'
 #' @references
-#' Rae, A. (2009). From spatial interaction data to spatial interaction information? Geovisualisation and spatial structures of migration from the 2001 UK census. Computers, Environment and Urban Systems, 33(3), 161–178. doi:10.1016/j.compenvurbsys.2009.01.007
+#' Rae, A. (2009). From spatial interaction data to spatial interaction information? Geovisualisation and spatial structures of migration from the 2001 UK census. Computers, Environment and Urban Systems, 33(3). doi:10.1016/j.compenvurbsys.2009.01.007
 #'
 #' @examples
 #' library(sp) # enable spatial objects
@@ -66,19 +66,27 @@ gFlow2line <- function(flow, zones){
 #' created using the \code{\link{gFlow2line}} function.
 #'
 #' @examples
-#' library(sp)
-#' v = data(package = "stplanr")
-#' data(list = v$results[,3])
-#' plot(l)
+#' library(rgdal)
+#' data(flowlines) # load demo flowlines dataset
+#' flowlines_84 <- sp::spTransform(flowlines, CRS("+init=epsg:4326"))
+#' plot(flowlines_84)
 #'
 #' \dontrun{
-#' # see http://www.cyclestreets.net/api/
+#' # set cyclestreets api key: see http://www.cyclestreets.net/api/
 #' cckey <- "f3fe3d078ac34737" # example key (not real)
-#' routes_fast <- gLines2CyclePath(l)
-#' routes_slow <- gLines2CyclePath(l, "quietest")
+#' # plot save the routes
+#' routes_fast <- gLines2CyclePath(flowlines_84)
+#' routes_slow <- gLines2CyclePath(flowlines_84, "quietest")
 #' }
-#' plot(routes_fast, col = "red", add = TRUE) # previously saved from l
-#' plot(routes_slow, col = "green", add = TRUE)
+#' data(routes_fast, routes_slow) # load routes
+#' lines(routes_fast, col = "red")
+#' lines(routes_slow, col = "green")
+#' # Plot for a single line to compare 'fastest' and 'quietest' route
+#' n = 19
+#' plot(flowlines_84[n,])
+#' lines(routes_fast[n,], col = "red")
+#' lines(routes_slow[n,], col = "green")
+
 gLines2CyclePath <- function(l, plan = "fastest"){
   if(!Sys.getenv('CYCLESTREET') == ""){
     cckey <- Sys.getenv('CYCLESTREET')
