+++ R/distance-decay-functions.R
@@ -1,68 +1,94 @@
 #' log-linear distance-decay function
 #'
-#' @section Details:
-#' Implementation of the formula used to fit distance to probability of
-#'  travelling by a given mode by Iacono et al. (2010)
+#' 'log-linear' distance decay function with two parameters:
+#' implementation of the formula used to fit distance to probability of
+#' travelling by a given mode by Iacono et al. (2010).
 #'
-#'  @seealso \code{\link{dd_logsqr}}
+#' @param x A positive vector representing distances (often in km)
+#' @param a Positive number (usually between 0 and 1) representing
+#'   the intercept of the distance decay curve with the y axis, when x = 0.
+#' @param b1 Number representing the 'beta' distance decay parameter.
+#' Larger negative values make the initial decay steeper.
+#'
 #' @references
 #' Iacono, M., Krizek, K. J. and El-Geneidy, A. (2010).
-#' Measuring non-motorized accessibility: issues, alternatives, and execution. Journal of Transport Geography, 18(1), 133–140. doi:10.1016/j.jtrangeo.2009.02.002
-#' @examples
-#' d <- 0:10 # vector of distances
-#' params <- c(0, -0.002, -0.15, -3)
-#' (res <- dd_iac(d, c(0, -0.002, -0.15, -3)))
-#' plot(d, res)
+#' Measuring non-motorized accessibility: issues, alternatives, and execution. Journal of Transport Geography, 18(1). doi:10.1016/j.jtrangeo.2009.02.002
 #'
-#' d = seq(0, 50, 0.1)
-#' plot(d, dd_iac(d))
-#' lines(d, dd_iac(d, b = 0.15))
-dd_iac <- function(d, a = 0.3, b = 0.2){
-  a * exp(-b * d)
+#' @examples
+#' x <- 0:10 # vector of distances
+#' a = 0.3 # default alpha value
+#' b = 0.2 # default beta value
+#' (res <- dd_loglin(x, a, b))
+#' plot(x, res)
+#'
+#' x = seq(0, 50, 0.1)
+#' plot(x, dd_loglin(x))
+#' lines(x, dd_loglin(x, a = 0.1, b1 = 0.15))
+dd_loglin <- function(x, a = 0.3, b1 = -0.2){
+  a * exp(b1 * x)
 }
-dd_iac(1:10)
 
 #' log-linear-square-root distance decay
 #'
-#' @section Details:
+#' @param x A positive vector representing distances (often in km)
+#' @param a Positive number (usually between 0 and 1) representing
+#'   the intercept of the distance decay curve with the y axis, when x = 0.
+#' @param b1 Number representing the 'beta' distance decay parameter.
+#' Higher values make the decay steeper.
+#' @param b2 Number representing the second 'beta' distance decay parameter.
+#' Higher values make the decay steeper.
+#'
 #' log-linear distance decay (Iacono et al., 2010) with an additional
-#' square-root term to make the curve much more flexible. The additional
-#' term (c in the function below) describes the short-term response:
-#' positive c values result in a unimodal distance decay with a peak after d = 0;
-#' negative c values lead to rapid decay in cycling under 2 km.
+#' square-root term to make the curve more flexible. The additional
+#' term (b2 in the function below) describes the short-term response:
[…]
