+++ man/age_recat.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2 (4.1.0): do not edit by hand
+% Please edit documentation in R/recat-NTS.R
+\name{age_recat}
+\alias{age_recat}
+\alias{age_recat2}
+\alias{disab_recat}
+\title{Recategorises UK Census age bands.}
+\usage{
+age_recat(a)
+
+age_recat2(a)
+
+disab_recat(a)
+}
+\arguments{
+\item{a}{The age variable supplied by the UK's NTS.}
+}
+\value{
+A factor with new age categories.
+}
+\description{
+\code{age_recat}, \code{age_recat2} and \code{disab_recat}
+are used to re-categorise the National Travel Survey age
+band and disability variables for comparison with other datasets.
+}
+\examples{
+age_recat("70 + years")
+}
+
