+++ README.Rmd
@@ -5,7 +5,12 @@ output:
     keep_md: yes
 ---
 
-This is an R package for sustainable transport planning in R (stplanr).
+```{r, echo=FALSE}
+library(stplanr)
+```
+
+
+This is a package for sustainable transport planning in R (stplanr).
 
 It brings together a range of tools for transport planning practitioners and
 researchers to better understand transport systems and inform policy.
@@ -15,6 +20,38 @@ The initial work on the project was funded by the Department of Transport
 as part of the National Propensity to Cycle Tool
 ([NPCT](http://www.cedar.iph.cam.ac.uk/research/modelling/npct-tool/)) project to
 identify where bicycle paths are most urgently needed.
+
+## Key functions
+
+Square data frames representing flows between origins and destinations
+must be combined with geo-referenced zones or points to generate meaningful
+analyses and visualisations of flows. **stplanr** facilitates this with 
+`gFlow2Line()`, which takes flow and geographical data as inputs and
+outputs a `SpatialLinesDataFrame`:
+
+```{r, results='hide', message=FALSE}
+library(stplanr)
+library(sp)
+data(cents, flow)
+flow[1:3, 1:3] # typical form of flow data
+cents[1:3,] # points representing origins and destinations
+```
+
+These datasets can be combined as follows:
+
+```{r}
+travel_network <- gFlow2line(flow = flow, zones = cents)
+w <- flow$All / max(flow$All) *10
+plot(travel_network, lwd = w)
+```
+
+The package can also allocate flows to the travel network, for example through
+a link to the [CycleStreets.net API](https://www.cyclestreets.net/api/):
+
+```{r, results='hide', message=FALSE}
+example("gLines2CyclePath")
+```
+
 
 ## Installation
 
@@ -30,8 +67,11 @@ We aim to make this package well-documented to make it easy to use.
 R's internal help functions will help here:
 
 ```{r}
-library(stplanr)
-?dd_iac # get help on an stplanr function
+?gLines2CyclePath # get help on an stplanr function
 lsf.str("package:stplanr", all = TRUE)
 ```
+
+Any questions?
+
+Email me on rob00 x at gmail dot com!
 
