### 2015-04-16 #10 @Robinlovelace
@sckott OK, I'm down to only 3 warnings from 7.

Any ideas how to remove the remaining 3.

Also check the updated README:
https://github.com/Robinlovelace/stplanr

What else do I need in there for this to be 'ropensci compliant'?
Thanks for the support, more to come.
