### 2015-04-17 e823a20
Added Barry as package author
