### 2015-04-17 a123927
Further updates to DESCRIPTION
