### 2015-04-17 ccb2c49
Add gOverlines functions
