### 2015-04-17 8e45b7a
Documentation updates
