+++ man/islines.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2 (4.1.0): do not edit by hand
+% Please edit documentation in R/gOverline.R
+\name{islines}
+\alias{islines}
+\title{Do the intersections between two geometries create lines?}
+\usage{
+islines(g1, g2)
+}
+\description{
+This is a function required in \code{\link{gOverline()}}. It identifies
+whether sets of lines overlap (beyond shared points) or
+not.
+}
+\examples{
+data(routes_fast)
+rnet <- gOverline(routes_fast[c(2, 3, 22),], attr = "length")
+r1 <- routes_fast[2,]
+r2 <- routes_fast[3,]
+r3 <- routes_fast[22,]
+plot(rnet)
+lines(r3, col = "red") # line without overlaps
+islines(r1, r2)
+islines(r1, r3)
+islines(r2, r3)
+}
+
