+++ DESCRIPTION
@@ -4,13 +4,11 @@ Title: Sustainable transport planning with R
 Version: 0.0.0.9005
 Date: 2015-03-10
 Authors@R: c(
-    person("Robin", "Lovelace", ,"rob00x@gmail.com", role = c("aut", "cre")),
-    person("Barry", "Rowlingson", role = c("aut"), comment = "Author of gOverline"),
+    person("Robin", "Lovelace", email = "rob00x@gmail.com", role = c("aut", "cre")),
+    person("Barry", "Rowlingson", role = c("aut"), comment = "Author of gOverline")
     )
 Maintainer: <rob00x@gmail.com>
-Description: Provides functionality for transport planning, including creation
-of origin-destination flow lines, route allocation and aggregation of routes
-to an on-road travel network.
+Description: Provides functionality for transport planning, including creation of origin-destination flow lines, route allocation and aggregation of routes to an on-road travel network.
 Depends: R (>= 3.1)
 License: MIT
 BugReports: https://github.com/robinlovelace/stplanr/issues
