### 2015-04-17 #10 @sckott
> WARNING '::' or ':::' import not declared from: ‘sp’  'library' or 'require' call not declared from: ‘sp’

* Looks like  you are using the `sp` package but you didn't put it into your DESCRIPTION file to load
* Looking at your NAMESPACE file, I'd suggest not using `exportPattern("^[[:alpha:]]+")` and instead using roxygen `@export` to explicitly declare what functions you want exported to users, and which not exported - also, using `@import` tags (e.g., `@import sp`) will put the appropriate entry in your NAMESPACE file for packages to import

> Namespaces in Imports field not imported from:   ‘dplyr’ ‘RCurl’ ‘rgdal’ ‘rgeos’   All declared Imports should be used.

* See above comment. Some of these packages are in your DESCRIPTION file, but some are not. R check found that you either A) use one of those packages in yours but it's not imported (in the DESCRIPTION and NAMESPACE files), or B) you don't use one of those packages in your package, but it's listed to import

> WARNING LaTeX errors when creating PDF version. This typically indicates Rd problems.

* Hard to say what's wrong here, other than that something with Rd files.

> WARNING '::' or ':::' import not declared from: ‘sp’ 'library' or 'require' call not declared from: ‘sp’

* You need to add sp to package imports

Also, I'm getting a bunch more warnings locally running check on your package.
