### 2015-04-17 #10 @Robinlovelace
@sckott Yes, it's because I just added [gOverline()](https://github.com/Robinlovelace/stplanr/blob/master/R/gOverline.R) and related functions which have caused additional warnings - there's been a lot of interest in the package of late and I expect more such functions, created by additional authors, to be added soon. Will debug based on your suggestions.
