### 2015-04-17 b5a53bf
Updated documentation to debug check() warnings
