+++ R/gOverline.R
@@ -1,8 +1,11 @@
 #' Do the intersections between two geometries create lines?
 #'
-#' This is a function required in \code{\link{gOverline()}}. It identifies
+#' This is a function required in \code{\link{gOverline}}. It identifies
 #' whether sets of lines overlap (beyond shared points) or
 #' not.
+#'
+#' @param g1 A SpatialLinesDataFrame
+#' @param g2 A SpatialLinesDataFrame
 #'
 #' @examples
 #' data(routes_fast)
@@ -17,7 +20,7 @@
 #' islines(r2, r3)
 islines <- function(g1, g2){
   ## return TRUE if geometries intersect as lines, not points
-  inherits(gIntersection(g1,g2), "SpatialLines")
+  inherits(rgeos::gIntersection(g1,g2), "SpatialLines")
 }
 
 #' Function to split overlapping SpatialLines into segments
@@ -26,24 +29,41 @@ islines <- function(g1, g2){
 #' Each new Lines object is the aggregate of a single number
 #' of aggregated lines.
 #'
+#' @param sl SpatialLinesDataFrame with overlapping Lines to split by
+#' number of overlapping features.
+#'
 #' @examples
+#' library(sp)
 #' data(routes_fast)
 #' rsec <- gSection(routes_fast)
 #' plot(rsec)
 #' length(rsec)
 #' set.seed(5)
 #' sel <- sample(length(rsec), 20)
-#' # plot(rsec[sel,], col = "red", add = T, lwd = 3) # overlapping lines
+#' # plot(rsec[sel,], col = "red", add = TRUE, lwd = 3) # overlapping lines
 gSection <- function(sl){
   ## union and merge and disaggregate to make a
   ## set of non-overlapping line segments
   sp::disaggregate(rgeos::gLineMerge(rgeos::gUnion(sl,sl)))
 }
 
+#' Label SpatialLinesDataFrame objects
+#'
+#' This function adds labels to lines plotted using base graphics. Largely
+#' for illustrative purposes, not designed for publication-quality
+#' graphics.
+#'
+#' @param sldf A SpatialLinesDataFrame with overlapping elements
+#' @param attr A text string corresponding to a named variable in \code{sldf}
+#'
+#' @author Barry Rowlingson
+#'
+#' @seealso \code{\link{gOverline}}
 lineLabels <- function(sldf, attr){
-  text(coordinates(gCentroid(sldf,byid=TRUE)),labels=sldf[[attr]])
+  text(sp::coordinates(
+    rgeos::gCentroid(sldf, byid = TRUE)
+    ), labels = sldf[[attr]])
 }
-
 
 #' Convert series of overlapping lines into a route network
 #'
@@ -53,23 +73,32 @@ lineLabels <- function(sldf, attr){
 #'
 #' @param sldf A SpatialLinesDataFrame with overlapping elements
 #' @param attr A text string corresponding to a named variable in \code{sldf}
+#' @param fun The function used to aggregate the grouped values (default: sum)
 #'
 #' @author Barry Rowlingson
 #'
+#' @references
+#'
+#' Rowlingson, B (2015). Overlaying lines and aggregating their values for
+#'  overlapping segments. Reproducible question from
+#'  \url{http://gis.stackexchange.com}. See \url{http://gis.stackexchange.com/questions/139681/overlaying-lines-and-aggregating-their-values-for-overlapping-segments}.
+#'
 #' @examples
 #' data(routes_fast)
+#' data(cents)
 #' rnet <- gOverline(sldf = routes_fast[1:7,], attr = "length")
 #' plot(rnet)
 #' points(cents)
 #' lineLabels(sldf = rnet, "length")
[…]
