+++ man/lineLabels.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2 (4.1.0): do not edit by hand
+% Please edit documentation in R/gOverline.R
+\name{lineLabels}
+\alias{lineLabels}
+\title{Label SpatialLinesDataFrame objects}
+\usage{
+lineLabels(sldf, attr)
+}
+\arguments{
+\item{sldf}{A SpatialLinesDataFrame with overlapping elements}
+
+\item{attr}{A text string corresponding to a named variable in \code{sldf}}
+}
+\description{
+This function adds labels to lines plotted using base graphics. Largely
+for illustrative purposes, not designed for publication-quality
+graphics.
+}
+\author{
+Barry Rowlingson
+}
+\seealso{
+\code{\link{gOverline}}
+}
+
