+++ DESCRIPTION
@@ -13,12 +13,13 @@ Depends: R (>= 3.1)
 License: MIT
 BugReports: https://github.com/robinlovelace/stplanr/issues
 Suggests:
+    dplyr,
     ggmap,
     tidyr,
     GISTools
 Imports:
     jsonlite,
-    dplyr,
-    rgdal,
     maptools,
-    rgeos
+    rgdal,
+    rgeos,
+    sp
