### 2015-04-17 #10 @Robinlovelace
@sckott Good news: it's now down to just *2 warnings* on my system after the latest commit:

```
WARNING
'library' or 'require' call not declared from: ‘sp’
See the information on DESCRIPTION files in the chapter ‘Creating R
packages’ of the ‘Writing R Extensions’ manual.
```

and

WARNING
LaTeX errors when creating PDF version.
This typically indicates Rd problems.
```

Not sure how to find exactly where the problems are emerging for either one. Any ideas?

Robin
