### 2015-04-17 #10 @sckott
@Robinlovelace still getting more errors than you are, what version of R are you using?

Also, looks like you are still using `exportPattern("^[[:alpha:]]+")`
