### 2015-04-18 2750fb3
Updates for R 3.2.0
