### 2015-04-18 #10 @Robinlovelace
@sckott I was using `3.1.2`. Now upgraded to `3.2.0`, had to change a few things to avoid errors. Looking at what I should put in the `NAMESPACE` now. Any pointers, other than the excellent resource on R Packages ([Wickham 2015](http://r-pkgs.had.co.nz/namespace.html))?
