### 2015-04-18 2c45cfd
Added NAMESPACE with Roxygen2 comments
