+++ man/writeGeoJSON.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.0): do not edit by hand
+% Generated by roxygen2 (4.1.1): do not edit by hand
 % Please edit documentation in R/geo-functions.R
 \name{writeGeoJSON}
 \alias{writeGeoJSON}
@@ -12,6 +12,8 @@ writeGeoJSON(x, filename)
 \item{filename}{File name of the output geojson}
 }
 \description{
-Provides a user-friendly wrapper for rgdal::writeOGR().
+Provides a user-friendly wrapper for rgdal::writeOGR(). Note,
+\code{geojson_write} from the geojsonio package
+provides the same functionality \url{https://github.com/ropensci/geojsonio}.
 }
 
