+++ man/gSection.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.0): do not edit by hand
+% Generated by roxygen2 (4.1.1): do not edit by hand
 % Please edit documentation in R/gOverline.R
 \name{gSection}
 \alias{gSection}
@@ -16,7 +16,6 @@ Each new Lines object is the aggregate of a single number
 of aggregated lines.
 }
 \examples{
-library(sp)
 data(routes_fast)
 rsec <- gSection(routes_fast)
 plot(rsec)
