+++ man/gFlow2line.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.0): do not edit by hand
+% Generated by roxygen2 (4.1.1): do not edit by hand
 % Please edit documentation in R/gFlow.R
 \name{gFlow2line}
 \alias{gFlow2line}
@@ -27,7 +27,6 @@ centroids. This can be tricky to plot and link-up with geographical data.
 This function makes the task easier.
 }
 \examples{
-library(sp) # enable spatial objects
 data(flow) # load data frame of od flows between zones
 data(cents) # load centroids data
 newflowlines <- gFlow2line(flow = flow, zones = cents)
