+++ NAMESPACE
@@ -1,11 +1,15 @@
 # Generated by roxygen2 (4.1.1): do not edit by hand
 
+export(age_recat)
+export(age_recat2)
 export(dd_logcub)
 export(dd_loglin)
 export(dd_logsqrt)
+export(disab_recat)
 export(gFlow2line)
 export(gLines2CyclePath)
 export(gMapshape)
 export(gOverline)
 export(gSection)
 export(islines)
+export(lineLabels)
