### 2015-04-18 f6a34bf
Regenerated NAMESPACE
