+++ NAMESPACE
@@ -1,4 +1,4 @@
-# Generated by roxygen2 (4.1.1): do not edit by hand
+# Generated by roxygen2 (4.1.0): do not edit by hand
 
 export(age_recat)
 export(age_recat2)
@@ -13,3 +13,4 @@ export(gOverline)
 export(gSection)
 export(islines)
 export(lineLabels)
+import(sp)
