+++ man/islines.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2 (4.1.0): do not edit by hand
 % Please edit documentation in R/gOverline.R
 \name{islines}
 \alias{islines}
@@ -17,6 +17,7 @@ whether sets of lines overlap (beyond shared points) or
 not.
 }
 \examples{
+\dontrun{
 data(routes_fast)
 rnet <- gOverline(routes_fast[c(2, 3, 22),], attr = "length")
 r1 <- routes_fast[2,]
@@ -27,5 +28,6 @@ lines(r3, col = "red") # line without overlaps
 islines(r1, r2)
 islines(r1, r3)
 islines(r2, r3)
+}
 }
 
