+++ man/gSection.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2 (4.1.0): do not edit by hand
 % Please edit documentation in R/gOverline.R
 \name{gSection}
 \alias{gSection}
@@ -16,6 +16,7 @@ Each new Lines object is the aggregate of a single number
 of aggregated lines.
 }
 \examples{
+\dontrun{
 data(routes_fast)
 rsec <- gSection(routes_fast)
 plot(rsec)
@@ -23,5 +24,6 @@ length(rsec)
 set.seed(5)
 sel <- sample(length(rsec), 20)
 # plot(rsec[sel,], col = "red", add = TRUE, lwd = 3) # overlapping lines
+}
 }
 
