+++ man/gFlow2line.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2 (4.1.0): do not edit by hand
 % Please edit documentation in R/gFlow.R
 \name{gFlow2line}
 \alias{gFlow2line}
@@ -27,11 +27,13 @@ centroids. This can be tricky to plot and link-up with geographical data.
 This function makes the task easier.
 }
 \examples{
+\dontrun{
 data(flow) # load data frame of od flows between zones
 data(cents) # load centroids data
 newflowlines <- gFlow2line(flow = flow, zones = cents)
 plot(cents)
 lines(newflowlines)
+}
 }
 \references{
 Rae, A. (2009). From spatial interaction data to spatial interaction information? Geovisualisation and spatial structures of migration from the 2001 UK census. Computers, Environment and Urban Systems, 33(3). doi:10.1016/j.compenvurbsys.2009.01.007
