### 2015-04-18 705e2da
Fixed *recat functions
