+++ R/recat-NTS.R
@@ -1,13 +1,23 @@
-#' Recategorises UK Census age bands.
+#' Recategorise UK National Travel Survey factor variables
 #'
 #' \code{age_recat}, \code{age_recat2} and \code{disab_recat}
-#' are used to re-categorise the National Travel Survey age
+#' are used to re-categorise the UK National Travel Survey (NTS) age
 #' band and disability variables for comparison with other datasets.
+#' See \url{http://discover.ukdataservice.ac.uk/series/?sn=2000037}
+#' for more on accessing the NTS.
 #'
-#' @param a The age variable supplied by the UK's NTS.
-#' @return A factor with new age categories.
+#' See Crawford and Lovelace (2015) for an application of these functions.
+#'
+#' @param a The factor variable supplied by the UK's NTS
+#' @return A factor with new categories for analysis
+#'
+#' @references
+#'
+#' Crawford, F., and Lovelace, R. (2015). The benefits of getting England cycling. Retrieved from http://www.ctc.org.uk/news/20150120-research-shows-growth-cycling-worth-%C2%BC-trillion-england%E2%80%99s-economy
+#'
 #' @export
 #' @examples
+#'
 #'
 #' age_recat("70 + years")
 age_recat <- function(a){
@@ -32,25 +42,25 @@ age_recat <- function(a){
 #' @rdname age_recat
 #' @export
 age_recat2 <- function(a){
-  c2 <- factor(rep(NA, length(c)), levels = c("0-16", "17-20", "21-29", "30-39","40-49",
+  a2 <- factor(rep(NA, length(a)), levels = c("0-16", "17-20", "21-29", "30-39","40-49",
     "50-59", "60-69", "70+"))
-  c2[c == "Less than 1 year" | c == "1 - 2 years" | c == "3 - 4 years"| c == "5 - 10 years" | c == "11 - 15 years"| c == "16 years"] <- "0-16"
-  c2[c == "17 years" | c == "18 years" | c == "19 years"| c == "20 years"] <- "17-20"
-  c2[c == "21 - 25 years" | c == "26 - 29 years"] <- "21-29"
-  c2[c == "30 - 39 years"] <- "30-39"
-  c2[c == "40 - 49 years"] <- "40-49"
-  c2[c == "50 - 59 years"] <- "50-59"
-  c2[c == "60 - 64 years" | c == "65 - 69 years"] <- "60-69"
-  c2[c == "70 - 74 years" | c == "75 - 79 years" | c == "80 - 84 years"| c == "85 years +"] <- "70+" # all people over 70...
-  c2
+  a2[a == "Less than 1 year" | a == "1 - 2 years" | a == "3 - 4 years"| a == "5 - 10 years" | a == "11 - 15 years"| a == "16 years"] <- "0-16"
+  a2[a == "17 years" | a == "18 years" | a == "19 years"| a == "20 years"] <- "17-20"
+  a2[a == "21 - 25 years" | a == "26 - 29 years"] <- "21-29"
+  a2[a == "30 - 39 years"] <- "30-39"
+  a2[a == "40 - 49 years"] <- "40-49"
+  a2[a == "50 - 59 years"] <- "50-59"
+  a2[a == "60 - 64 years" | a == "65 - 69 years"] <- "60-69"
+  a2[a == "70 - 74 years" | a == "75 - 79 years" | a == "80 - 84 years"| a == "85 years +"] <- "70+" # all people over 70...
+  a2
 }
 
 #' @rdname age_recat
 #' @export
 disab_recat <- function(a){
[…]
