### 2015-04-18 941e950
Updated documentation
