### 2015-04-19 11da1d4
Create new gOneway function
