### 2015-04-19 59fcb4c
Updated documentation
