+++ R/gOverline.R
@@ -9,15 +9,17 @@
 #' @export
 #' @examples \dontrun{
 #' data(routes_fast)
-#' rnet <- gOverline(routes_fast[c(2, 3, 22),], attr = "length")
+#' rnet <- gOverline(routes_fast[c(2, 3, 22),], attrib = "length")
 #' r1 <- routes_fast[2,]
 #' r2 <- routes_fast[3,]
 #' r3 <- routes_fast[22,]
+#' library(sp)
 #' plot(rnet)
 #' lines(r3, col = "red") # line without overlaps
 #' islines(r1, r2)
 #' islines(r1, r3)
 #' islines(r2, r3)
+#'
 #' }
 islines <- function(g1, g2){
   ## return TRUE if geometries intersect as lines, not points
@@ -36,16 +38,18 @@ islines <- function(g1, g2){
 #' @examples \dontrun{
 #' data(routes_fast)
 #' rsec <- gSection(routes_fast)
-#' plot(rsec)
+#' library(sp)
+#' plot(routes_fast)
+#' lines(rsec, col = "red", lwd = 3)
 #' length(rsec)
 #' set.seed(5)
 #' sel <- sample(length(rsec), 20)
-#' # plot(rsec[sel,], col = "red", add = TRUE, lwd = 3) # overlapping lines
+#' plot(rsec[sel,], col = "blue", add = TRUE, lwd = 3) # overlapping lines
 #' }
 gSection <- function(sl){
   ## union and merge and disaggregate to make a
   ## set of non-overlapping line segments
-  sp::disaggregate(rgeos::gLineMerge(rgeos::gUnion(sl,sl)))
+  sp::disaggregate(rgeos::gLineMerge(rgeos::gUnion(sl, sl)))
 }
 
 #' Label SpatialLinesDataFrame objects
@@ -55,16 +59,16 @@ gSection <- function(sl){
 #' graphics.
 #'
 #' @param sldf A SpatialLinesDataFrame with overlapping elements
-#' @param attr A text string corresponding to a named variable in \code{sldf}
+#' @param attrib A text string corresponding to a named variable in \code{sldf}
 #'
 #' @author Barry Rowlingson
 #'
 #' @seealso \code{\link{gOverline}}
 #' @export
-lineLabels <- function(sldf, attr){
+lineLabels <- function(sldf, attrib){
   text(sp::coordinates(
     rgeos::gCentroid(sldf, byid = TRUE)
-    ), labels = sldf[[attr]])
+    ), labels = sldf[[attrib]])
 }
 
 #' Convert series of overlapping lines into a route network
@@ -74,8 +78,9 @@ lineLabels <- function(sldf, attr){
 #' and converts these into a single route network.
 #'
 #' @param sldf A SpatialLinesDataFrame with overlapping elements
-#' @param attr A text string corresponding to a named variable in \code{sldf}
+#' @param attrib A text string corresponding to a named variable in \code{sldf}
 #' @param fun The function used to aggregate the grouped values (default: sum)
+#' @param na.zero Sets whether aggregated values with a value of zero are removed.
 #'
 #' @author Barry Rowlingson
 #' @references
@@ -87,13 +92,29 @@ lineLabels <- function(sldf, attr){
 #' @examples \dontrun{
 #' data(routes_fast)
 #' data(cents)
-#' rnet <- gOverline(sldf = routes_fast[1:7,], attr = "length")
+#' rnet <- gOverline(sldf = routes_fast[1:7,], attrib = "length")
+#' library(sp) # needed for plotting spatial objects
 #' plot(rnet)
 #' points(cents)
 #' lineLabels(sldf = rnet, "length")
[…]
