+++ man/lineLabels.Rd
@@ -1,15 +1,15 @@
-% Generated by roxygen2 (4.1.0): do not edit by hand
+% Generated by roxygen2 (4.1.1): do not edit by hand
 % Please edit documentation in R/gOverline.R
 \name{lineLabels}
 \alias{lineLabels}
 \title{Label SpatialLinesDataFrame objects}
 \usage{
-lineLabels(sldf, attr)
+lineLabels(sldf, attrib)
 }
 \arguments{
 \item{sldf}{A SpatialLinesDataFrame with overlapping elements}
 
-\item{attr}{A text string corresponding to a named variable in \code{sldf}}
+\item{attrib}{A text string corresponding to a named variable in \code{sldf}}
 }
 \description{
 This function adds labels to lines plotted using base graphics. Largely
