+++ man/gOneway.Rd
@@ -1,0 +1,43 @@
+% Generated by roxygen2 (4.1.1): do not edit by hand
+% Please edit documentation in R/gOneway.R
+\name{gOneway}
+\alias{gOneway}
+\title{Aggregate flows so they become non-directional}
+\usage{
+gOneway(x, attrib)
+}
+\arguments{
+\item{x}{A SpatialLinesDataFrame}
+
+\item{attrib}{A text string containing the name of the line's attribute to aggregate}
+}
+\value{
+\code{gOneway} outputs a SpatialLinesDataFrame with single lines
+and user-selected attribute values that have been aggregated. Only lines
+with a distance (i.e. not intra-zone flows) are included
+}
+\description{
+Flow data often contains movement in two directions: from point A to point B
+and then from B to A. This can be problematic for transport planning, because
+the magnitude of flow along a route can be masked by flows the other direction.
+If only the largest flow in either direction is captured in an analysis, for
+example, the true extent of travel will by heavily under-estimated for
+OD pairs which have similar amounts of travel in both directions.
+Flows in both direction are often represented by overlapping lines with
+identical geometries (see \code{\link{flowlines}}) which can be confusing
+for users and are difficult to plot.
+}
+\details{
+This function aggregates directional flows into non-directional flows,
+potentially halving the number of lines objects and reducing the number
+of overlapping lines to zero.
+}
+\examples{
+library(sp)
+data("flowlines")
+plot(flowlines)
+singlelines <- gOneway(flowlines, attrib = "All")
+nrow(singlelines)
+sum(singlelines$All)
+}
+
