+++ man/dd_loglin.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.0): do not edit by hand
+% Generated by roxygen2 (4.1.1): do not edit by hand
 % Please edit documentation in R/distance-decay-functions.R
 \name{dd_loglin}
 \alias{dd_loglin}
