### 2015-04-19 #10 @sckott
@Robinlovelace nice, checked again with your most recent version and got no warnings
