### 2015-04-19 174ea77
Updated function
