### 2015-04-19 c416e77
Add gOneway* functions
