+++ R/gOverline.R
@@ -6,7 +6,9 @@
 #'
 #' @param g1 A SpatialLinesDataFrame
 #' @param g2 A SpatialLinesDataFrame
+#' @family Functions for manipulating flow data
 #' @export
+#'
 #' @examples \dontrun{
 #' data(routes_fast)
 #' rnet <- gOverline(routes_fast[c(2, 3, 22),], attrib = "length")
@@ -34,6 +36,7 @@ islines <- function(g1, g2){
 #'
 #' @param sl SpatialLinesDataFrame with overlapping Lines to split by
 #' number of overlapping features.
+#' @family Functions for manipulating flow data
 #' @export
 #' @examples \dontrun{
 #' data(routes_fast)
@@ -63,7 +66,7 @@ gSection <- function(sl){
 #'
 #' @author Barry Rowlingson
 #'
-#' @seealso \code{\link{gOverline}}
+#' @family Functions for manipulating flow data
 #' @export
 lineLabels <- function(sldf, attrib){
   text(sp::coordinates(
@@ -88,6 +91,7 @@ lineLabels <- function(sldf, attrib){
 #' Rowlingson, B (2015). Overlaying lines and aggregating their values for
 #'  overlapping segments. Reproducible question from
 #'  \url{http://gis.stackexchange.com}. See \url{http://gis.stackexchange.com/questions/139681/overlaying-lines-and-aggregating-their-values-for-overlapping-segments}.
+#' @family Functions for manipulating flow data
 #' @export
 #' @examples \dontrun{
 #' data(routes_fast)
@@ -141,3 +145,132 @@ gOverline <- function(sldf, attrib, fun = sum, na.zero = FALSE){
 
   sldf
 }
+
+#' Aggregate flows so they become non-directional (by geometry - the slow way)
+#'
+#' Flow data often contains movement in two directions: from point A to point B
+#' and then from B to A. This can be problematic for transport planning, because
+#' the magnitude of flow along a route can be masked by flows the other direction.
+#' If only the largest flow in either direction is captured in an analysis, for
+#' example, the true extent of travel will by heavily under-estimated for
+#' OD pairs which have similar amounts of travel in both directions.
+#' Flows in both direction are often represented by overlapping lines with
+#' identical geometries (see \code{\link{flowlines}}) which can be confusing
+#' for users and are difficult to plot.
+#'
+#' This function aggregates directional flows into non-directional flows,
+#' potentially halving the number of lines objects and reducing the number
+#' of overlapping lines to zero.
+#'
+#' @param x A SpatialLinesDataFrame
+#' @param attrib A text string containing the name of the line's attribute to
+#' aggregate or a numeric vector of the columns to be aggregated
+#'
+#' @return \code{gOnewaygeo} outputs a SpatialLinesDataFrame with single lines
+#' and user-selected attribute values that have been aggregated. Only lines
+#' with a distance (i.e. not intra-zone flows) are included
+#' @family Functions for manipulating flow data
+#' @export
+#' @examples
+#' library(sp)
+#' data("flowlines")
+#' plot(flowlines)
+#' singlelines <- gOnewaygeo(flowlines, attrib = 3:14)
+#' plot(singlelines, lwd = 3, col = "red")
+#' lines(singlelines) # check we've got the right lines
+#' sum(singlelines$All)
+#' nrow(singlelines)
+gOnewaygeo <- function(x, attrib){
+  geq <- rgeos::gEquals(x, x, byid = T)
+  sel1 <- !duplicated(geq) # repeated rows
+  sel2 <- rowSums(geq) > 1 # all features with an overlap
+  sel3 <- rgeos::gLength(x, byid = TRUE) > 0 # remove points
[…]
