+++ NAMESPACE
@@ -9,7 +9,8 @@ export(disab_recat)
 export(gFlow2line)
 export(gLines2CyclePath)
 export(gMapshape)
-export(gOneway)
+export(gOnewaygeo)
+export(gOnewayid)
 export(gOverline)
 export(gSection)
 export(islines)
