### 2015-04-19 1106d3f
Updated documentation
