+++ man/flow.Rd
@@ -34,9 +34,5 @@ use_data(routes_fast)
 use_data(routes_slow)
 }
 }
-\seealso{
-Other res: \code{\link{flowlines}};
-  \code{\link{routes_fast}}; \code{\link{routes_slow}}
-}
 \keyword{datasets}
 
