+++ man/lineLabels.Rd
@@ -20,6 +20,8 @@ graphics.
 Barry Rowlingson
 }
 \seealso{
-\code{\link{gOverline}}
+Other Functions for manipulating flow data: \code{\link{gOnewaygeo}};
+  \code{\link{gOnewayid}}; \code{\link{gOverline}};
+  \code{\link{gSection}}; \code{\link{islines}}
 }
 
