+++ man/routes_slow.Rd
@@ -11,9 +11,5 @@ data(routes_slow)
 \description{
 SpatialLinesDataFrame of commuter flows on the travel network
 }
-\seealso{
-Other res: \code{\link{flowlines}}; \code{\link{flow}};
-  \code{\link{routes_fast}}
-}
 \keyword{datasets}
 
