### 2015-04-20 9ca7ab3
Fix gOneway* bug
