### 2015-04-22 #10 @Robinlovelace
@sckott I'm still getting a warning with the LaTeX. Trying to debug it now: http://stackoverflow.com/questions/29791500/how-to-identify-latex-errors-in-rd-r-help-files

 Next stage: tests with [testthat](http://r-pkgs.had.co.nz/tests.html). Any advice for the tests? Can I use the packages own data to run the tests (e.g. `cents` in stplanr)?
