### 2015-04-22 e5b9f22
Updated latest log and DESCRIPTION
