### 2015-04-22 #10 @Robinlovelace
I've also updated the first comment of this thread to reflect recent improvements in stplanr.
