### 2015-04-22 #10 @karthik
@Robinlovelace Do you have a vignette? If not, just ignore the latex warnings. Those are unrelated to the package itself.
