### 2015-04-22 #10 @karthik
> Can I use the packages own data to run the tests (e.g. cents in stplanr)?
Yes!
Re Travis, I'd suggest starting here and we can help from there. http://r-pkgs.had.co.nz/check.html#travis
