### 2015-04-22 #10 @jennybc
Re: using package's own data, here's a tiny example:

https://github.com/jennybc/gapminder/blob/master/tests/testthat/test-gapminder.R

`gapminder` is just a data package and has only this one test. But `data/gapminder.rda` is where the data lives within the package and I can just refer to the `gapminder` data.frame w/o further fuss in my tests. You should be able to do same with your data objects `cents`, `flow`, etc.

Re: Travis. If you are using `devtools`, there is a `use_travis()` function to set that process in motion. Note: The development version from GitHub has been updated to produce the new, simpler `.travis.yml` file, as illustrated in the link above. The released CRAN version produces an older, hairier file, FYI.
