### 2015-04-24 03a9603
Add gClip function
