+++ NAMESPACE
@@ -6,6 +6,7 @@ export(dd_logcub)
 export(dd_loglin)
 export(dd_logsqrt)
 export(disab_recat)
+export(gClip)
 export(gFlow2line)
 export(gLines2CyclePath)
 export(gMapshape)
