### 2015-05-07 a9ebd03
Added initial travis yml
