+++ DESCRIPTION
@@ -7,7 +7,7 @@ Authors@R: c(
     person("Robin", "Lovelace", email = "rob00x@gmail.com", role = c("aut", "cre")),
     person("Barry", "Rowlingson", role = c("aut"), comment = "Author of gOverline")
     )
-Description: Provides functionality for transport planning, including creation of origin-destination flow lines, route allocation and aggregation of routes to an on-road travel network.
+Description: Functionality and data acces for transport planning, including origin-destination analysis, route allocation and modelling travel patterns.
 License: MIT + file LICENSE
 BugReports: https://github.com/robinlovelace/stplanr/issues
 LazyData: yes
