### 2015-05-08 93b8322
Added libgdal-dev dependency
