### 2015-05-08 a5daae1
Added R binary package
