### 2015-05-08 1660ca9
Remove external dependencies
