+++ .travis.yml
@@ -5,8 +5,6 @@ r_binary_packages:
   - rgdal
 
 apt_packages:
-- libhdf5-serial-dev
-- libgdal-dev
 - texlive-science
 - latex-xcolor
 
