### 2015-05-08 b621365
Remove external dependencies
