### 2015-05-08 855d453
New vignette, other updates
