+++ README.Rmd
@@ -5,6 +5,8 @@ output:
     keep_md: yes
 ---
 
+[[Build Status](https://travis-ci.org/Robinlovelace/stplanr.svg?branch=master)](https://travis-ci.org/Robinlovelace/stplanr)
+
 ```{r, echo=FALSE}
 library(stplanr)
 ```
@@ -78,4 +80,14 @@ lsf.str("package:stplanr", all = TRUE)
 Any questions?
 
 Email me on rob00 x at gmail dot com!
+
+```{r, echo=FALSE}
+md <- readLines("README.md")
+sel <- grep(pattern = "Build", x = md)
+bsl <- md[sel]
+bsl <- gsub(pattern = "\\[\\[Build Status]", replacement = "\\[!\\[Build Status]", x = bsl)
+md[sel] <- bsl
+writeLines(text = md, con = "README.md")
+```
+
 
