+++ DESCRIPTION
@@ -11,16 +11,18 @@ Description: Functionality and data acces for transport planning, including orig
 License: MIT + file LICENSE
 BugReports: https://github.com/robinlovelace/stplanr/issues
 LazyData: yes
-Depends:
+Depends: 
     R (>= 2.10)
-Imports:
+Imports: 
     jsonlite,
     maptools,
     raster,
     rgdal,
     rgeos,
     sp
-Suggests:
+Suggests: 
     GISTools,
     igraph,
-    osmar
+    osmar,
+    knitr
+VignetteBuilder: knitr
