### 2015-05-08 #10 @Robinlovelace
Thanks for the information @jennybc, I've now added my first unit test (for `gFlow2Line`) and it seems to work! I will do the same for the datasets if needs be, but need to prioritise tests for the most important things...

@sckott I'm on the R Packages course taught by Colin Gillespie in Newcastle now. I've added continuous testing with Travis and (after installing r-gdal from the system, not R) it builds successfully! I've also added a vignette and a demo.

Next step that should make the package even more attractive from an rOpenSci perspective is the addition of tools to automatically download and query OSM data: empowering people via data access. When this is implemented, can we move forward by assigning an editor etc?

Finally, by adding a blank line in `demo/00Index` (randomly!) and adding **sp** as a dependency (it really is) I've just eliminated all warnings on my system. Wahey!
