### 2015-05-08 583198b
Added demo, test and updated DESCRIPTION
