+++ demo/route-analysis.R
@@ -1,0 +1,14 @@
+data("cents", package = "stplanr")
+library(sp)
+plot(cents)
+data("flowlines")
+lines(flowlines)
+data("routes_fast")
+lines(routes_fast, col = "red")
+library(rgeos)
+library(rgdal)
+lgb <- spTransform(flowlines, CRSobj = CRS("+init=epsg:27700"))
+length1 <- gLength(lgb, byid = T)
+length2 <- routes_fast@data$length
+plot(length1, length2, xlab = "Euclidean distance", ylab = "Route distance")
+abline(a = 0, b = 1)
