### 2015-05-08 3477fa9
Latest log and vignette updates
