### 2015-05-08 97f8f9c
Simplify travis instructions
