+++ .travis.yml
@@ -3,3 +3,8 @@ sudo: true
 
 r_binary_packages:
   - rgdal
+
+notifications:
+    email:
+      on_success: change
+      on_failure: change
