### 2015-05-08 948ba7c
Made sp a dependency and associated changes
