### 2015-05-08 4104d2c
Typo correction and package help page
