+++ R/stplanr-package.R
@@ -1,0 +1,11 @@
+#' The stplanr package
+#'
+#' The stplanr package provides functions to access and analyse data for transportation research, including origin-destination analysis, route allocation and modelling travel patterns.
+#'
+#' @name stplanr-package
+#' @aliases stplanr stplanr-package
+#' @docType package
+#' @author \email{rob00x@@gmail.com}
+#' @keywords package
+#' @seealso \url{https://github.com/Robinlovelace/stplanr}
+NULL
