+++ man/stplanr-package.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2 (4.1.1): do not edit by hand
+% Please edit documentation in R/stplanr-package.R
+\docType{package}
+\name{stplanr-package}
+\alias{stplanr}
+\alias{stplanr-package}
+\title{The stplanr package}
+\description{
+The stplanr package provides functions to access and analyse data for transportation research, including origin-destination analysis, route allocation and modelling travel patterns.
+}
+\author{
+\email{rob00x@gmail.com}
+}
+\seealso{
+\url{https://github.com/Robinlovelace/stplanr}
+}
+\keyword{package}
+
