### 2015-05-08 906efe5
Update to vignette, removal of igraph suggest
