+++ R/gFlow.R
@@ -47,7 +47,7 @@ gFlow2line <- function(flow, zones){
 #' You will need to have a cyclestreets api key for this function to work, e.g.:
 #'
 #' \code{
-#' mytoken <- readLines("~/Dropbox/dotfiles/cyclestreets-api-key-rl")
+#' mytoken <- "f3fe3d078ac34737" # warning: not a real key
 #' }
 #'
 #' \code{
