### 2015-05-08 #10 @sckott
Nice progress!  yes, adding OSM data access would fit better into our wheel-house
