+++ vignettes/refs.bib
@@ -1,0 +1,66 @@
+Automatically generated by Mendeley Desktop 1.13.8
+Any changes to this file will be lost if it is regenerated by Mendeley.
+
+BibTeX export options can be customized via Options -> BibTeX in Mendeley Desktop
+
+@article{Waddell2002,
+author = {Waddell, Paul},
+issn = {0194-4363},
+journal = {Journal of the American Planning Association},
+number = {3},
+pages = {297--314},
+publisher = {Taylor \& Francis},
+title = {{UrbanSim: Modeling urban development for land use, transportation, and environmental planning}},
+volume = {68},
+year = {2002}
+}
+@article{Martinez2013,
+abstract = {This paper tries to break new ground in how distance-decay relationships are modelled in accessibility and transport demand studies and does it based on an innovative approach to empirical data collection on psychological perceptions of distance in relation with activities located in space and a new aggregate distance-decay function. This new approach improves on the quality of the representation of spatial interaction effects on transport demand modelling studies that commonly rely on generic curves barely confronted with empirical data. We compare the level of fit of the proposed curve with other distance-decay functions mentioned in the literature and used in practice and draw relevant conclusions on the proper model specification. © 2012 Elsevier Ltd.},
+author = {Mart\'{\i}nez, L. Miguel and Viegas, Jos\'{e} Manuel},
+doi = {10.1016/j.jtrangeo.2012.08.018},
+file = {:media/robin/data/Copy/1Lit/2013/Mart\'{\i}nez, Viegas - 2013 - A new approach to modelling distance-decay functions for accessibility assessment in transport studies.pdf:pdf},
+issn = {09666923},
+journal = {Journal of Transport Geography},
+keywords = {Accessibility studies,Distance-decay functions,Spatial impedance functions,Spatial interaction models},
+pages = {87--96},
+title = {{A new approach to modelling distance-decay functions for accessibility assessment in transport studies}},
+volume = {26},
+year = {2013}
+}
+@article{Iacono2010,
+author = {Iacono, Michael and Krizek, Kevin J. and El-Geneidy, Ahmed},
+doi = {10.1016/j.jtrangeo.2009.02.002},
+file = {:media/robin/data/Copy/1Lit/2010/Iacono, Krizek, El-Geneidy - 2010 - Measuring non-motorized accessibility issues, alternatives, and execution.pdf:pdf},
+issn = {09666923},
+journal = {Journal of Transport Geography},
+month = jan,
+number = {1},
+pages = {133--140},
+publisher = {Elsevier Ltd},
+title = {{Measuring non-motorized accessibility: issues, alternatives, and execution}},
+url = {http://linkinghub.elsevier.com/retrieve/pii/S0966692309000210},
+volume = {18},
+year = {2010}
+}
+@article{Kahle2013,
+author = {Kahle, D and Wickham, Hadley},
+file = {:media/robin/data/Copy/1Lit/2013/Kahle, Wickham - 2013 - ggmap Spatial Visualization with ggplot2.pdf:pdf},
+journal = {The R Journal},
+pages = {144--161},
+title = {{ggmap: Spatial Visualization with ggplot2}},
+url = {http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.375.8693\&rep=rep1\&type=pdf},
+volume = {5},
+year = {2013}
+}
+@incollection{Cheshire2015,
+author = {Cheshire, James and Lovelace, Robin},
+booktitle = {Geocomputation},
+chapter = {1},
+editor = {Brunsdon, Chris and Singleton, Alex},
[…]
