+++ vignettes/introducing-stplanr.Rmd
@@ -2,6 +2,7 @@
 title: "Introducing stplanr"
 author: "Robin Lovelace"
 date: "`r Sys.Date()`"
+bibliography: refs.bib
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Introducing stplanr}
@@ -24,7 +25,8 @@ package name ("sustainable transport planning with R")
 refers to both the focus on active travel
 and the aim for longevity and accessibility in the software.
 Transport planning is notoriously reliant on 'black boxes' and
-the same applies to scientific research into transport systems.
+the same applies to scientific research into transport systems
+[@Waddell2002].
 
 **stplanr** seeks to address these issues. After the package
 has been installed (see the package's
@@ -107,7 +109,10 @@ routes to the transport network using the
 an API key for the function to work):
 
 ```{r}
-if(nchar(Sys.getenv('CYCLESTREET')) == 16){
+x = try(GET("https://github.com"), silent = T) # test the connection
+offline <- grepl("Error", x)
+
+if(nchar(Sys.getenv('CYCLESTREET')) == 16 & !offline){
   rf <- gLines2CyclePath(l = flowlines)
 } else{
   data("routes_fast")
@@ -118,17 +123,138 @@ plot(l)
 lines(rf, col = "red")
 ```
 
-What this does is fire requests to CycleStreets.net.
+The above code sends requests to CycleStreets.net and saves the results.
 By changing the `plan` argument of `gLines2Cyclepath()`, we can download
 routes that are more suitable for people prioritising speed, quietness or
-a balance between speed and quietness. To illustrate the difference between
-different types of route, the figure below shows 'fast' and 'quiet' routes
-allocated to the transport network in Manchester (the package 'leaflet'
-was used to generate the basemap).
+a balance between speed and quietness. The plots of `l` and `rf` illustrate
+how the spatial classes that the flow data are converted to are easy to
+visualise. Using such data objects, generated by **stplanr**, in combination
+with additional packages for creating maps in R
+[e.g. see @Cheshire2015; @Kahle2013], will further enhance the utility of
+the package's functions. Route-allocated flows are also useful for 
+analysis.
+
+## Analysis of origin-destination flow data
+
+Route-allocated lines allow estimation of *route distance* and 
+*cirquity* (route distance divided by Euclidean distance).
+These variables can help model the rate of flow between origins and
+destination, as illustrated in the next section. The code below demonstrates
+how objects generated by **stplanr** can be used to estimate route distance.
+
+```{r}
+library(rgdal) # required for spTransform
+lgb <- spTransform(l, CRSobj = CRS("+init=epsg:27700"))
+l$d_euclidean <- rgeos::gLength(lgb, byid = T)
+l$d_fastroute <- routes_fast@data$length
+plot(l$d_euclidean, l$d_fastroute,
+  xlab = "Euclidean distance", ylab = "Route distance")
+abline(a = 0, b = 1)
+abline(a = 0, b = 1.2, col = "green")
+abline(a = 0, b = 1.5, col = "red")
+```
+
+The plot illustrates the expected strong correlation between
+Euclidean and fastest route distance. However, some OD pairs
+have a proportionally higher route distance than others.
+This is illustrated
+by distance from the black line in the above plot, which represents circuity. 
+
+The difference between
+different types of route is shown in the figure below.
+'Fast' and 'quiet' routes (red and green, respectively)
+are clearly allocated to the transport network
+(the package 'leaflet' was used to generate the basemap).
 
 ![](figures/fast-slow-manc.png)
 
-## Developing models of transport
[…]
