### 2015-05-09 8d77222
More vignette updates and --as-cran check arg
