### 2015-05-09 f3deb37
Further updates to vignette, remove figure from vignette folder
