+++ vignettes/introducing-stplanr.Rmd
@@ -9,6 +9,15 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+```{r, include=FALSE}
+library(stplanr)
+library(sp) # needed for geographical objects
+library(ggmap) # needed for plotting data
+library(rgdal) # required for spTransform
+library(RCurl)
+```
+
+
 ## Introduction
 
 Common tasks in transportation research and planning include:
@@ -34,9 +43,7 @@ it can be loaded in with `library()`:
 
 ```{r, message=FALSE}
 library(stplanr)
-library(sp) # needed for geographical objects
 ```
-
 
 ## Accessing and converting data
 
@@ -108,14 +115,18 @@ routes to the transport network using the
 an API key for the function to work):
 
 ```{r}
-x = try(GET("https://github.com"), silent = T) # test the connection
+# test the connection
+x = try(download.file("http://www.google.com", "test.html"), silent = T) 
 offline <- grepl("Error", x)
 
 if(nchar(Sys.getenv('CYCLESTREET')) == 16 & !offline){
   rf <- gLines2CyclePath(l = flowlines)
+  rq <- gLines2CyclePath(l = flowlines, plan = "quietest")
 } else{
   data("routes_fast")
+  data("routes_slow")
   rf <- routes_fast
+  rq <- routes_slow
 }
 
 plot(l)
@@ -142,7 +153,6 @@ destination, as illustrated in the next section. The code below demonstrates
 how objects generated by **stplanr** can be used to estimate route distance.
 
 ```{r}
-library(rgdal) # required for spTransform
 lgb <- spTransform(l, CRSobj = CRS("+init=epsg:27700"))
 l$d_euclidean <- rgeos::gLength(lgb, byid = T)
 l$d_fastroute <- routes_fast@data$length
@@ -165,7 +175,25 @@ different types of route is shown in the figure below.
 are clearly allocated to the transport network
 (the package 'leaflet' was used to generate the basemap).
 
-![](figures/fast-slow-manc.png)
+
+```{r, echo=FALSE, warning=FALSE, message=FALSE}
+if(offline){
+  print("Figure not shown: requires internet connection")
+} else {
+  n <- 30
+  lf <- fortify(l[n,])
+  rff <- fortify(rf[n,])
+  rqf <- fortify(rq[n,])
+  bb <- bbox(rf[n,])
+  bb <- make_bbox(lon = bb[1,], lat = bb[2,], f = 0.5)
+  ggmap(ggmap = get_map(location = bb)) +
+    geom_path(aes(x = long, y = lat), data = lf) + 
+    geom_path(aes(x = long, y = lat), data = rqf, colour = "red") + 
+    geom_path(aes(x = long, y = lat), data = rff, colour = "green") +
+    theme_nothing()
+}
+```
+
 
 ## Developing models of travel behaviour
 
