### 2015-05-09 147633a
Updated dependencies for Travis build
