### 2015-05-09 6169a8a
Remove testthat Travis dependency
