### 2015-05-09 1f1dd5e
remove rmarkdown dependency
