### 2015-05-09 897b0b1
rm r_packages
