### 2015-05-12 53af4eb
Added CRAN badge, fixed DESCRIPTION, debugging
