+++ DESCRIPTION
@@ -24,5 +24,5 @@ Suggests:
     osmar,
     knitr,
     testthat,
-    ggmap,
+    ggmap
 VignetteBuilder: knitr
