### 2015-05-31 3de3f5b
Make gOverline more robust
