### 2015-06-04 eaea18f
Added bbox scale and beginnings of osm route network dl function
