+++ R/osm.R
@@ -1,0 +1,121 @@
+# OSM tools - for downloading osm data
+
+# data("flowlines")
+# bb <- sp::bbox(flowlines)
+# bbox_2_overpass <- function(bb, type = "("){
+#   if(type == "["){
+#     paste0("[bbox=", paste(bb, collapse = ","), "]")
+#   } else{
+#     paste0("(", paste(bb, collapse = ","), ")")
+#   }
+# }
+#
+# get_over(bb)
+# cb <- rgeos::gCentroid(flowlines)
+# cb2 <- center_bbox(coordinates(cb)[1], coordinates(cb)[2], 200, 200)
+# lways <- get_osm(cb2, src)
+#
+# get_over <- function(bb){
+#   uri = "http://overpass-api.de/api/interpreter?data="
+#   if(length(dim(bb)) != 1)
+#     bb <- bbox_2_overpass(bb)
+#   dreq <- '[timeout:25];'
+#   nreq <- '(node["highway"]'
+#   wreq <- ';way["highway"]'
+#   rreq <- ';relation["highway"]'
+#   freq <- ';);out body;>;out skel qt;'
+#   req <- paste0(dreq, nreq, bb, wreq, bb, rreq, bb, freq)
+#   req
+#   req <- RCurl::curlEscape(req)
+#   req <- paste0(uri, req)
+#   downloader::download(req, destfile = "/tmp/output.osm")
+# }
+#
+# req <- '[out:json][timeout:25];(node["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053);way["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053);relation["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053););out body;>;out skel qt;'
+#
+# library(downloader)
+# downloader::download(get_over(bb), "here.osm")
+#
+#
+# osm_to_geojson("/tmp/output.osm")
+# library(sp)
+# osm_dat <- rgdal::readOGR("/tmp/output.geojson", "OGRGeoJSON") # fail
+# osm_dat <- geojsonio::geojson_read("/tmp/output.geojson") # fail
+# plot(osm_dat)
+# geojsonio::topojson_read()
+# osm_dat[[3]][4]
+#
+#
+# class(osm_dat)
+# plot(osm_dat)
+#
+# osm_to_geojson <- function(file){
+#   newname <- gsub(pattern = ".osm", replacement = ".geojson", file)
+#   mess <- paste0("osmtogeojson ", file, " > ", newname)
+#   system(mess)
+# }
+#
+# downloader::download(url = URLencode('http://overpass-api.de/api/interpreter?data=[out:json][timeout:25];(node["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053);way["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053);relation["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053););out body;>;out skel qt;)'), destfile = "/tmp/test.json")
+#
[…]
