### 2015-06-09 bb42b63
Updated gClip so it returns @data
