### 2015-06-09 017ed7f
Update gLines2CyclePath and add dplyr to suggests
