### 2015-06-18 63fca84
Added new vignette on flow data
