+++ vignettes/Transport-r-flows.bib
@@ -1,0 +1,54 @@
+Automatically generated by Mendeley Desktop 1.13.8
+Any changes to this file will be lost if it is regenerated by Mendeley.
+
+BibTeX export options can be customized via Options -> BibTeX in Mendeley Desktop
+
+@article{Rae2009,
+author = {Rae, Alasdair},
+doi = {10.1016/j.compenvurbsys.2009.01.007},
+file = {:media/robin/data/Copy/1Lit/2009/Rae - 2009 - From spatial interaction data to spatial interaction information Geovisualisation and spatial structures of migration from.pdf:pdf},
+isbn = {0198-9715},
+issn = {01989715},
+journal = {Computers, Environment and Urban Systems},
+keywords = {Flow mapping,Folder - GraphVisualization,GIS,Geovisualisation,Migration,Spatial interaction data,Urska supplied paper,cartography,demography,flow,jtc,movement,network-analysis,social network,spatial,spatial interaction data,spatial-networks,visualization},
+mendeley-tags = {Flow mapping,Folder - GraphVisualization,GIS,Geovisualisation,Migration,Spatial interaction data,Urska supplied paper,cartography,demography,flow,jtc,movement,network-analysis,social network,spatial,spatial-networks,visualization},
+month = may,
+number = {3},
+pages = {161--178},
+publisher = {Elsevier Ltd},
+title = {{From spatial interaction data to spatial interaction information? Geovisualisation and spatial structures of migration from the 2001 UK census}},
+url = {http://linkinghub.elsevier.com/retrieve/pii/S019897150900009X},
+volume = {33},
+year = {2009}
+}
+@book{kabacoff2011r,
+author = {Kabacoff, Robert},
+publisher = {Manning Publications Co.},
+title = {{R in Action}},
+year = {2011}
+}
+@article{Fotheringham1981,
+author = {Fotheringham, AS},
+file = {:media/robin/data/Copy/1Lit/1981/Fotheringham - 1981 - Spatial structure and distance-decay parameters.pdf:pdf},
+journal = {Annals of the Association of American Geographers},
+number = {3},
+pages = {425--436},
+title = {{Spatial structure and distance-decay parameters}},
+url = {http://www.tandfonline.com/doi/abs/10.1111/j.1467-8306.1981.tb01367.x},
+volume = {71},
+year = {1981}
+}
+@article{lovelace2014introduction,
[…]
