### 2015-06-21 ffe5f8f
Create route_cyclestreet() - next step: pull-out info
