+++ NAMESPACE
@@ -17,4 +17,5 @@ export(gOverline)
 export(gSection)
 export(islines)
 export(lineLabels)
+export(route_cyclestreet)
 import(sp)
