### 2015-06-22 5db8a6a
Edits to route_cyclestreet() documentation
