+++ man/route_cyclestreet.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/cyclestreet_route.R
+% Please edit documentation in R/routes.R
 \name{route_cyclestreet}
 \alias{route_cyclestreet}
 \title{Plan a single route with CycleStreets.net}
@@ -51,7 +51,12 @@ plot(rq_hfd)
 # Plan a 'balanced' route from Pedaller's Arms in Leeds
 to the University of Leeds
 
-route_cyclestreet(from = "Pedaller's Arms, Leeds", to = "University of Leeds", plan = "balanced")
+rb_pa <- route_cyclestreet(from = "Pedaller's Arms, Leeds", to = "University of Leeds", plan = "balanced")
+
+library(leaflet) # requires the leaflet CRAN package
+# display the route on an interactive map
+leaflet() \%>\% addTiles() \%>\%
+  addPolylines(data = rb_pa)
 
 # Plan a route between two lat/lon pairs in the UK
 
