### 2015-07-06 0c3e23c
Update cc function for modified v2 api
