### 2015-07-26 4695619
Added graphhopper routing functionality
