+++ R/osm.R
@@ -1,12 +1,12 @@
-# OSM tools - for downloading osm data
+# OSM tools - for downloading osm data - this is work in progress
 
 # data("flowlines")
 # bb <- sp::bbox(flowlines)
 # bbox_2_overpass <- function(bb, type = "("){
 #   if(type == "["){
-#     paste0("[bbox=", paste(bb, collapse = ","), "]")
+#     paste0("[bbox=", paste(bb[c(2,1,4,3)], collapse = ","), "]")
 #   } else{
-#     paste0("(", paste(bb, collapse = ","), ")")
+#     paste0("(", paste(bb[c(2,1,4,3)], collapse = ","), ")")
 #   }
 # }
 #
@@ -31,8 +31,6 @@
 #   downloader::download(req, destfile = "/tmp/output.osm")
 # }
 #
-# req <- '[out:json][timeout:25];(node["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053);way["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053);relation["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053););out body;>;out skel qt;'
-#
 # library(downloader)
 # downloader::download(get_over(bb), "here.osm")
 #
@@ -40,7 +38,10 @@
 # osm_to_geojson("/tmp/output.osm")
 # library(sp)
 # osm_dat <- rgdal::readOGR("/tmp/output.geojson", "OGRGeoJSON") # fail
-# osm_dat <- geojsonio::geojson_read("/tmp/output.geojson") # fail
+# osm_dat <- geojsonio::geojson_read("/tmp/output.geojson")
+# osm_dat$features
+#
+# osm_paths <- osm_dat$features
 # plot(osm_dat)
 # geojsonio::topojson_read()
 # osm_dat[[3]][4]
@@ -54,68 +55,3 @@
 #   mess <- paste0("osmtogeojson ", file, " > ", newname)
 #   system(mess)
 # }
-#
-# downloader::download(url = URLencode('http://overpass-api.de/api/interpreter?data=[out:json][timeout:25];(node["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053);way["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053);relation["highway"](41.88539386041403,12.485854625701903,41.89461068146964,12.498128414154053););out body;>;out skel qt;)'), destfile = "/tmp/test.json")
-#
-# http://overpass-api.de/api/interpreter?data=%5Bout%3Ajson%5D%5Btimeout%3A25%5D%3B%28node%5B%22highway%22%5D%2841%2E88539386041403%2C12%2E485854625701903%2C41%2E89461068146964%2C12%2E498128414154053%29%3Bway%5B%22highway%22%5D%2841%2E88539386041403%2C12%2E485854625701903%2C41%2E89461068146964%2C12%2E498128414154053%29%3Brelation%5B%22highway%22%5D%2841%2E88539386041403%2C12%2E485854625701903%2C41%2E89461068146964%2C12%2E498128414154053%29%3B%29%3Bout%20body%3B%3E%3Bout%20skel%20qt%3B%0A
-#
-# http://overpass-api.de/api/interpreter?
[…]
