### 2015-07-26 97d324a
Graphhopper namespace
