+++ NAMESPACE
@@ -18,4 +18,5 @@ export(gSection)
 export(islines)
 export(lineLabels)
 export(route_cyclestreet)
+export(route_graphhopper)
 import(sp)
