### 2015-08-04 ca7c7cc
Fix gLines2CyclePath code and improve route_cyclestreet
