### 2015-09-24 1fa59be
Update description of gOverline
