### 2015-09-26 49e400a
Create simple line2route function
