+++ man/line2route.Rd
@@ -1,0 +1,44 @@
+% Generated by roxygen2 (4.1.1): do not edit by hand
+% Please edit documentation in R/gFlow.R
+\name{line2route}
+\alias{line2route}
+\title{Convert straight SpatialLinesDataFrame from flow data into routes}
+\usage{
+line2route(ldf, ...)
+}
+\arguments{
+\item{l}{A SpatialLinesDataFrame or data.frame of coordinates produced by
+\code{\link{line2df}}}
+}
+\description{
+Convert straight SpatialLinesDataFrame from flow data into routes
+}
+\section{Details}{
+
+
+See \code{\link{route_cyclestreet}} and other route functions for details
+}
+\examples{
+library(rgdal)
+data(flowlines) # load demo flowlines dataset
+flowlines_84 <- sp::spTransform(flowlines, CRS("+init=epsg:4326"))
+fl <- flowlines_84[rgeos::gLength(flowlines_84, byid = T) > 0,]
+plot(fl)
+
+\dontrun{
+
+routes_f <- line2route(fl)
+routes_s <- line2route(fl, plan = "quietest", silent = T)
+}
+if(!exists("routes_f")){
+  data(routes_fast, routes_s) # load routes
+}
+lines(routes_f, col = "red")
+lines(routes_s, col = "green")
+# Plot for a single line to compare 'fastest' and 'quietest' route
+n = 18
+plot(fl[n,])
+lines(routes_f[n,], col = "red")
+lines(routes_s[n,], col = "green")
+}
+
