### 2015-10-11 9e16798
Fix test issue with T vs TRUE
