+++ R/gFlow.R
@@ -184,9 +184,11 @@ gLines2CyclePath <- function(l, plan = "fastest"){
 #'
 #' @examples
 #' library(rgdal)
+#' library(sp)
 #' data(flowlines) # load demo flowlines dataset
 #' flowlines_84 <- sp::spTransform(flowlines, CRS("+init=epsg:4326"))
-#' fl <- flowlines_84[rgeos::gLength(flowlines_84, byid = T) > 0,]
+#' sel <- rgeos::gLength(flowlines_84, byid = TRUE) > 0
+#' fl <- flowlines_84[sel,]
 #' plot(fl)
 #'
 #' \dontrun{
@@ -226,3 +228,4 @@ line2route <- function(ldf, ...){
   }
   rf
 }
+
