### 2015-10-11 37ecdb7
Update instructions for CycleStreet api key
