### 2015-10-11 6eb4da6
Added readTableBuilder function. Function imports ABS TableBuilder data and returns a dataframe.
