### 2015-10-12 c659743
Remove @family tag from gOverline functions
