+++ man/gLines2CyclePath.Rd
@@ -34,6 +34,18 @@ Sys.setenv(CYCLESTREET = mytoken) # see http://www.cyclestreets.net/api/
 \code{
 cckey <- Sys.getenv('CYCLESTREET')
 }
+
+or on Linux
+
+\code{
+line_to_add <- "export CYCLESTREET='f3fe3d078ac34737'"
+write(line_to_add, "~/.bashrc", append = TRUE)
+}
+
+Then restart so the envrionment variable is set and can be viewed with
+\code{
+$CYCLESTREET
+}
 }
 \examples{
 library(rgdal)
