+++ man/gOverline.Rd
@@ -55,9 +55,4 @@ Rowlingson, B (2015). Overlaying lines and aggregating their values for
  overlapping segments. Reproducible question from
  \url{http://gis.stackexchange.com}. See \url{http://gis.stackexchange.com/questions/139681/overlaying-lines-and-aggregating-their-values-for-overlapping-segments}.
 }
-\seealso{
-Other Functions for manipulating flow data: \code{\link{gOnewaygeo}};
-  \code{\link{gOnewayid}}; \code{\link{gSection}};
-  \code{\link{islines}}; \code{\link{lineLabels}}
-}
 
