+++ man/routes_slow.Rd
@@ -11,5 +11,9 @@ data(routes_slow)
 \description{
 SpatialLinesDataFrame of commuter flows on the travel network
 }
+\seealso{
+Other res: \code{\link{flowlines}}; \code{\link{flow}};
+  \code{\link{routes_fast}}
+}
 \keyword{datasets}
 
