### 2015-10-12 c4073e6
Remove superfluous 'require' call
Causes a warning in devtools::check()
