### 2015-10-12 4b3e664
Remove gLines2CyclePath, repo tidying
