+++ R/gFlow.R
@@ -64,123 +64,6 @@ line2df <- function(l){
   output
 }
 
-#' Convert straight SpatialLinesDataFrame from flow data into routes in the UK
-#' using the CycleStreets.net API.
-#'
-#' @section Details:
-#'
-#' You will need to have a cyclestreets api key for this function to work, e.g.:
-#'
-#' \code{
-#' mytoken <- "f3fe3d078ac34737" # warning: not a real key
-#' }
-#'
-#' \code{
-#' Sys.setenv(CYCLESTREET = mytoken) # see http://www.cyclestreets.net/api/
-#' }
-#'
-#' \code{
-#' cckey <- Sys.getenv('CYCLESTREET')
-#' }
-#'
-#' or on Linux
-#'
-#' \code{
-#' line_to_add <- "export CYCLESTREET='f3fe3d078ac34737'"
-#' write(line_to_add, "~/.bashrc", append = TRUE)
-#' }
-#'
-#' Then restart so the envrionment variable is set and can be viewed with
-#' \code{
-#' $CYCLESTREET
-#' }
-#'
-#' @param plan A text string. Must be either "balanced", "fastest" (default)
-#' or "quietest"
-#'
-#' @param l A SpatialLinesDataFrame object composed of straight lines. l may be
-#' created using \code{\link{od2line}}.
-#'
-#' @inheritParams route_cyclestreet
-#'
-#' @export
-#'
-#' @examples
-#' library(rgdal)
-#' data(flowlines) # load demo flowlines dataset
-#' flowlines_84 <- sp::spTransform(flowlines, CRS("+init=epsg:4326"))
-#' plot(flowlines_84)
-#'
-#' \dontrun{
-#' # set cyclestreets api key: see http://www.cyclestreets.net/api/
-#' cckey <- "f3fe3d078ac34737" # example key (not real)
-#' # plot save the routes
-#'
-#' routes_fast <- gLines2CyclePath(flowlines_84)
-#' routes_slow <- gLines2CyclePath(flowlines_84, "quietest")
-#' }
-#' if(!exists("routes_fast")){
-#'   data(routes_fast, routes_slow) # load routes
-#' }
-#' lines(routes_fast, col = "red")
-#' lines(routes_slow, col = "green")
-#' # Plot for a single line to compare 'fastest' and 'quietest' route
-#' n = 20
-#' plot(flowlines_84[n,])
-#' lines(routes_fast[n,], col = "red")
-#' lines(routes_slow[n,], col = "green")
-
-gLines2CyclePath <- function(l, plan = "fastest"){
-  if(!Sys.getenv('CYCLESTREET') == ""){
-    cckey <- Sys.getenv('CYCLESTREET')
-  }
-  if(is.null(cckey)){
-    stop("You must have a CycleStreets.net api key saved as 'cckey'")
-  }
-  coord_list <- lapply(slot(l, "lines"), function(x) lapply(slot(x, "Lines"),
-    function(y) slot(y, "coords")))
-  output <- vector("list", length(coord_list))
-  api_base <- sprintf("https://api.cyclestreets.net/v2/")
-  for(i in 1:length(output)){
-    from <- coord_list[[i]][[1]][1, ]
-    to <- coord_list[[i]][[1]][2, ]
-
-    if(identical(from, to)){
-      # if the length is zero...
-      route <-
-        sp::SpatialLines(list(Lines(list(Line(rbind(from, to))),
-          row.names(l[i,]))))
-      df <- data.frame(matrix(NA, ncol = 6))
-      names(df) <- c("plan", "start", "finish", "length", "time", "waypoint")
-      row.names(df) <- row.names(l[i,])
-      route <- SpatialLinesDataFrame(route, df)
-    }else{
-      # save lines that have a distance
[…]
