+++ R/data.r
@@ -84,8 +84,8 @@ NULL
 #' use_data(flowlines)
 #'
 #' # Convert flows to routes
-#' routes_fast <- gLines2CyclePath(l = flowlines, plan = "fastest")
-#' routes_slow <- gLines2CyclePath(l = flowlines, plan = "quietest")
+#' routes_fast <- line2route(l = flowlines, plan = "fastest")
+#' routes_slow <- line2route(l = flowlines, plan = "quietest")
 #'
 #' use_data(routes_fast)
 #' use_data(routes_slow)
