### 2015-10-12 9038b50
Update documentation and README
