+++ README.Rmd
@@ -74,8 +74,10 @@ R's internal help functions will help here:
 ```{r}
 ?route_cyclestreet # get help on an stplanr function
 if(Sys.info()["sysname"] == "Linux"){
+  # On Linux
   mytoken <- readLines("~/Dropbox/dotfiles/cyclestreets-api-key-rl")
 } else {
+  # Example on Windows
   mytoken <- readLines("file:///C:/Users/georl/Dropbox/dotfiles/cyclestreets-api-key-rl")
 }
 Sys.setenv(CYCLESTREET = mytoken)
