### 2015-10-12 #10 @Robinlovelace
Hi @sckott I think we're ready to move this forward again towards peer review. @richardellison has added functionality for reading in data from official Australian sources and the addition of `route_graphhopper()` means that we can now use this to access routed osm data from anywhere in the world. E.g. (once you have the graphhopper api key):

```
#' @examples
#'
#' \dontrun{
#' r <- route_graphhopper("Leeds", "Dublin", vehicle = "bike")
#'
#' library(leaflet)
#'
#' leaflet() %>% addTiles() %>% addPolylines(data = r)
#' }
```
![leeds-route](https://cloud.githubusercontent.com/assets/1825120/10425588/af30cc8a-70d0-11e5-8e62-b0509da41c92.png)
