### 2015-10-12 3478b04
Update graphhopper docs
