### 2015-10-12 2635cd7
Update documentation to remove undocumented data
Also add more detail on adding system environment
