+++ R/gFlow.R
@@ -79,28 +79,35 @@ line2df <- function(l){
 #'
 #' @examples
 #' library(rgdal)
-#' library(sp)
 #' data(flowlines) # load demo flowlines dataset
-#' flowlines_84 <- sp::spTransform(flowlines, CRS("+init=epsg:4326"))
-#' sel <- rgeos::gLength(flowlines_84, byid = TRUE) > 0
-#' fl <- flowlines_84[sel,]
-#' plot(fl)
+#' flowlines <- spTransform(flowlines, CRS("+init=epsg:27700"))
+#' flowlines <- flowlines[rgeos::gLength(flowlines, byid = TRUE) > 0,]
+#' flowlines <- spTransform(flowlines, CRS("+init=epsg:4326"))
+#' plot(flowlines)
 #'
 #' \dontrun{
+#' cckey <- readLines("~/Dropbox/dotfiles/cyclestreets-api-key-rl")
+#' Sys.setenv(CYCLESTREET = cckey)
+#' routes_fast <- line2route(l = flowlines, plan = "fastest")
+#' routes_slow <- line2route(l = flowlines, plan = "quietest", silent = TRUE)
+#' }
 #'
-#' routes_f <- line2route(fl)
-#' routes_s <- line2route(fl, plan = "quietest", silent = T)
+#' # Save the route data (uncomment if this changes)
+#' # devtools::use_data(routes_fast, overwrite = TRUE)
+#' # devtools::use_data(routes_slow, overwrite = TRUE)
+#'
+#' if(!exists("routes_fast")){
+#'   data(routes_fast, routes_slow) # load routes
 #' }
-#' if(!exists("routes_f")){
-#'   data(routes_fast, routes_s) # load routes
-#' }
-#' lines(routes_f, col = "red")
-#' lines(routes_s, col = "green")
+#'
+#' lines(routes_fast, col = "red")
+#' lines(routes_slow, col = "green")
+#'
 #' # Plot for a single line to compare 'fastest' and 'quietest' route
 #' n = 18
 #' plot(fl[n,])
-#' lines(routes_f[n,], col = "red")
-#' lines(routes_s[n,], col = "green")
+#' lines(routes_fast[n,], col = "red")
+#' lines(routes_slow[n,], col = "green")
 
 line2route <- function(ldf, ...){
   if(class(ldf) == "SpatialLinesDataFrame") ldf <- line2df(ldf)
