### 2015-10-12 d52a5cd
Update package to fix chcke warnings
