+++ man/line2df.Rd
@@ -13,7 +13,6 @@ line2df(l)
 Convert straight SpatialLinesDataFrame to a data.frame with from and to coords
 }
 \examples{
-library(rgdal)
 data(flowlines) # load demo flowlines dataset
 ldf <- line2df(flowlines)
 }
