+++ man/route_graphhopper.Rd
@@ -25,7 +25,9 @@ To use this function you will need to obtain an API key from
 \url{https://graphhopper.com/#directions-api}.
 It is assumed that you have set your api key as a system environment
 for security reasons (so you avoid typing the API key in your code).
-Do this with \code{GRAPHHOPPER}, e.g. by typing \code{Sys.setenv(GRAPHHOPPER = 'eccbf612-214e-437d-8b73-06bdf9e68731')}. (Note: key not real, use your own key.)
+Do this with \code{GRAPHHOPPER}, e.g. by typing \code{
+Sys.setenv(GRAPHHOPPER = 'eccbf612-214e-437d-8b73-06bdf9e6877d')
+}. (Note: key not real, use your own key.)
 }
 \details{
 The function returns a SpatialLinesDataFrame object.
@@ -36,7 +38,8 @@ your own API key:
 
 \code{obj <- jsonlite::fromJSON(url)}
 
-Where \code{url} is an example api request from \url{https://github.com/graphhopper/directions-api/blob/master/docs-routing.md}.
+Where \code{url} is an example api request from
+ \url{https://github.com/graphhopper/directions-api/blob/master/docs-routing.md}.
 }
 \examples{
 \dontrun{
