### 2015-10-12 14e2dbf
Update line2route documentation (plot error solution)
