+++ R/gFlow.R
@@ -104,6 +104,7 @@ line2df <- function(l){
 #'   data(routes_fast, routes_slow) # load routes
 #' }
 #'
+#' plot(flowlines)
 #' lines(routes_fast, col = "red")
 #' lines(routes_slow, col = "green")
 #'
