### 2015-10-12 #10 @Robinlovelace
Yes. I can polish up the vignette, make sure all is ship shape and create additional 'feature requests' this week. But yes I think now's a good time to get this 'on the road'.
