### 2015-10-13 d1cda08
Changes to try to fix errors on documentation and package build
