+++ man/flowlines.Rd
@@ -9,11 +9,8 @@
 data(flowlines)
 }
 \description{
-SpatialLinesDataFrame of commuter flows
-}
-\seealso{
-Other res: \code{\link{flow}}; \code{\link{routes_fast}};
-  \code{\link{routes_slow}}
+Flow data after conversion to a spatial format
+with \code{\link{od2line}} (see \code{\link{flow}}).
 }
 \keyword{datasets}
 
