### 2015-10-16 #10 @sckott
Super nice package! nice work all.

* Use `@importFrom` whenever possible. Right now you have `import(openxlsx)` and `import(sp)` in your `NAMESPACE` file. Just import the functions you need. Same for other pkg deps.
* You need to import packages explicitly that you use. You have a bunch of packages in your `DESCRIPTION` file that are not imported. Looks like your namespacing function calls (e.g., `sp::SpatialLinesDataFrame`), which is fine if the user happens to have `sp` installed, but will fail if the user doesn't. Ideally, import only functions you use, and place those imports somewhere in your package (I like to put them all in one file, the pacakge level man file, e.g., `stplanr-package.R`)
* On check I saw `Rd file 'route_cyclestreet.Rd'` - make sure line widths for roxygen docs aren't more than 100 characters.
* It appears that `XML` package is not used at all? If not, remove from `DESCRIPTION` file.
* Tests: Pleae write tests to cover at least all the major functions before we accept. Use `testthat::skip_on_cran()` for any tests that do web requests, so that CRAN tests don't fail in case a service is temporarily down
* Title in `DESCRIPTION` file should be sentence case
* Use a `URL` field in your `DESCRIPTION` file: (e.g., `URL: https://github.com/ropensci/taxize`)
* Curious whether you really need `sp` in `Depends`, or if you could have in `Imports`. Fine if you need in depends, but if you don't absolutely need it there, best not to
* README:
    * At bottom of readme, you say to contact you via email. That's cool, but also direct to the Issue tracker in the repo
    * Point to instructions in readme for installing `rgdal` and `rgeos`, they are often painful for folks
* I think token's can be a bit easier for the user. Right now you have  e.g.,

    ```r
    if (!Sys.getenv('CYCLESTREET') == "") {
        cckey <- Sys.getenv('CYCLESTREET')
    }
    if(is.null(cckey)){
        stop("You must have a CycleStreets.net api key saved as 'cckey'")
    }
    ```

    However, if someone like me, is using this pkg for the first time, they may not have gotten a key yet, and therefore that first `if` statement will lead to not creating a `cckey`  variable, which then makes the next `if` statement fail as it's not found.  And I think you should tell them the name of the environment variable to set (`CYCLESTREET`), not the internal var name `cckey`. You could chang this up a bit so that the user is given a prompt if a key isn't found, though not necessary

* `gOverline` for me makes R crash. e.g., this `gOverline(routes_fast[c(2, 3, 22),], attrib = "length")`.  Does that work for you? Looks like it fails here for me (from the `gOverline` function):

    ```r
    slu = gSection(sl)
    #> Assertion failed: (!"should never be reached"), function itemsTree, file AbstractSTRtree.cpp, line 371.
    #> Abort trap: 6
    ```

* Throughtout package, make sure to have examples that will work for __anyone__. E.g,. for `readTableBuilder()` you have examples like `readTableBuilder('employmentSLA.csv')`, but the user most likely will not have that file. You can put a csv file in `inst/examples`, though make sure it's small-ish.
* Looks like for web requests you're doing are directly through `jsonlite::fromJSON()`. We strongly urge you to use `httr` to do web requests. It provides a better interface for doing web requests, fails better than jsonlite will, and allows users the ability to toggle curl options. Still good to parse the result with `jsonlite::fromJSON()` though. Let me know if you need help with this.
* Use a code of conduct, see `devtools::use_code_of_conduct()`
* We strongly prefer `snake_case` for function names https://github.com/ropensci/packaging_guide#funvar - Would you be up for changing function names?

I'm waiting to get api keys from CycleSteets and GraphHopper - and I'll have more feedback on those functions after I get keys
