### 2015-10-18 1912d17
Added catchment area functions based on straight-line distance, minor fixes for gClip function, rename readTableBuilder to snake case and update function to account for updated TableBuilder outputs.
