+++ R/catchmentArea.R
@@ -1,0 +1,223 @@
+#' Calculate catchment area and associated summary statistics.
+#'
+#' @section Details:
+#' Calculates the catchment area of a facility (e.g., cycle path) using
+#' straight-line distance as well as summary statistics from variables
+#' available in a SpatialPolygonsDataFrame with census tracts or other
+#' zones. Assumes that the frequency of the variable is evenly distributed
+#' throughout the zone. Returns a SpatialPolygonsDataFrame.
+#'
+#' @param polygonlayer A SpatialPolygonsDataFrame containing zones from which
+#' the summary statistics for the catchment variable will be calculated.
+#' Smaller polygons will increase the accuracy of the results.
+#' @param targetlayer A SpatialPolygonsDataFrame, SpatialLinesDataFrame,
+#' SpatialPointsDataFrame, SpatialPolygons, SpatialLines or SpatialPoints
+#' object containing the specifications of the facility for which the
+#' catchment area is being calculated. If the object contains more than one
+#' facility (e.g., multiple cycle paths) the aggregate catchment area will be
+#' calculated.
+#' @param calccols A vector of column names containing the variables in the
+#' polygonlayer to be used in the calculation of the summary statistics for
+#' the catchment area. If dissolve = FALSE, all other variables in the
+#' original SpatialPolygonsDataFrame for zones that fall partly or entirely
+#' within the catchment area will be included in the returned
+#' SpatialPolygonsDataFrame but will not be adjusted for the proportion within
+#' the catchment area.
+#' @param distance Defines the size of the catchment area as the distance
+#' around the targetlayer in the units of the projection
+#' (default = 500 metres)
+#' @param projection The proj4string used to define the projection to be used
+#' for calculating the catchment areas or a character string 'austalbers' to
+#' use the Australian Albers Equal Area projection. Ignored if the polygonlayer
+#' is projected in which case the targetlayer will be converted to the
+#' projection used by the polygonlayer. In all cases the resulting object will
+#' be reprojected to the original coordinate system and projection of the
+#' polygon layer. Default is an Albers Equal Area projection but for more
+#' reliable results should use a local projection (e.g., Australian Albers
+#' Equal Area project).
+#' @param retainAreaProportion Boolean value. If TRUE retains a variable in
+#' the resulting SpatialPolygonsDataFrame containing the proportion of the
+#' original area within the catchment area (Default = FALSE).
+#' @param dissolve Boolean value. If TRUE collapses the underlying zones
+#' within the catchment area into a single region with statistics for the
+#' whole catchment area.
+#' @export
+#' @examples \dontrun{
+#' calc_catchment(
+#'    polygonlayer = sa1income,
+#'    targetlayer = testcycleway,
+#'    calccols = c('Total'),
+#'    distance = 800,
+#'    projection = 'austalbers',
+#'    dissolve = TRUE
+#' )
+#' }
+calc_catchment <- function(
+  polygonlayer,
+  targetlayer,
+  calccols,
+  distance = 500,
+  projection = "+proj=aea +lat_1=90 +lat_2=-18.416667 +lat_0=0 +lon_0=10 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs",
+  retainAreaProportion = FALSE,
+  dissolve = FALSE
+){
+
+  # Define Named vector of known projection strings
+  knownprojs <- c(
[…]
