### 2015-10-18 071148e
Add rgeos to r_binary packages
