+++ .travis.yml
@@ -1,12 +1,9 @@
 language: r
 sudo: true
 
-before_install:
-  - sudo apt-get install libgdal-dev libproj-dev libgeos-dev
-
 r_binary_packages:
-  - colorspace
   - rgdal
+  - rgeos
 
 # r_packages:
 
