### 2015-10-18 c312cfa
remove rgeos as binary from .travis.yml
