+++ .travis.yml
@@ -1,4 +1,5 @@
 language: r
+sudo: true
 
 r_binary_packages:
   - rgdal
