### 2015-10-18 03e5b67
Add ggplot2 binary package to travis
