+++ .travis.yml
@@ -3,6 +3,7 @@ sudo: true
 
 r_binary_packages:
   - rgdal
+  - ggplot2
 
 # r_packages:
 
