### 2015-10-18 60c33de
Add dissolve = FALSE to @example
This fixes a WARNING from devtools::check()
