+++ R/catchmentArea.R
@@ -60,7 +60,7 @@ calc_catchment <- function(
   projection = "+proj=aea +lat_1=90 +lat_2=-18.416667 +lat_0=0 +lon_0=10 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs",
   retainAreaProportion = FALSE,
   dissolve = FALSE
-){
+  ){
 
   # Define Named vector of known projection strings
   knownprojs <- c(
@@ -182,7 +182,7 @@ calc_catchment <- function(
 #' original area within the catchment area (Default = FALSE).
 #' @param dissolve Boolean value. If TRUE collapses the underlying zones
 #' within the catchment area into a single region with statistics for the
-#' whole catchment area.
+#' whole catchment area. Default is FALSE.
 #' @export
 #' @examples \dontrun{
 #' calc_catchment_sum(
@@ -190,7 +190,8 @@ calc_catchment <- function(
 #'    targetlayer = testcycleway,
 #'    calccols = c('Total'),
 #'    distance = 800,
-#'    projection = 'austalbers'
+#'    projection = 'austalbers',
+#'    dissolve = FALSE
 #' )
 #' }
 calc_catchment_sum <- function(
