### 2015-10-18 2fff807
Bump version number, add tmap, remove GISTools deps
