### 2015-10-18 13e96c9
Add namespace to Line() function for check()
