### 2015-10-18 ac7a2f1
Add namespace to proj4string
