### 2015-10-18 401de45
Update stplanr package documentation
This is designed to fix this issue:
https://github.com/ropensci/onboarding/issues/10
"You need to import packages explicitly that you use."
