### 2015-10-18 e780258
Update README to link to gh issues
