### 2015-10-18 e4c623c
Update README to explain binary install of rgdal
