### 2015-10-18 5700324
Reference Code of Conduct in README, ignore in build
