### 2015-10-18 f4e28a9
Change case of title, update R version, remove XML
