### 2015-10-18 bf573eb
Remove imports() from NAMESPACE
