+++ NAMESPACE
@@ -23,5 +23,3 @@ export(od2line)
 export(read_table_builder)
 export(route_cyclestreet)
 export(route_graphhopper)
-import(openxlsx)
-import(sp)
