+++ NAMESPACE
@@ -23,3 +23,11 @@ export(od2line)
 export(read_table_builder)
 export(route_cyclestreet)
 export(route_graphhopper)
+import(rgdal)
+import(rgeos)
+import(sp)
+importFrom(ggmap,geocode)
+importFrom(jsonlite,fromJSON)
+importFrom(openxlsx,readWorkbook)
+importFrom(raster,crop)
+importFrom(raster,extent)
