### 2015-10-18 6271997
Prevent nchar > 100 in script
