### 2015-10-18 a8e5b4e
Delete redundant code
