+++ inst/wip.R
@@ -1,9 +1,0 @@
-# Work in progress
-
-# # Version1 api (works)
-# requestv1 <- "http://www.cyclestreets.net/api/journey.xml?key="
-# requestv1 <- paste0(requestv1, cckey)
-# requestv1 <- paste0(requestv1, "&itinerarypoints=", ft_string, "&plan=")
-# requestv1 <- paste0(requestv1, plan)
-#
-# obj2 <- XML::xmlTreeParse(requestv1)
