+++ man/stplanr-package.Rd
@@ -4,12 +4,22 @@
 \name{stplanr-package}
 \alias{stplanr}
 \alias{stplanr-package}
-\title{The stplanr package}
+\title{\strong{stplanr: Sustainable Transport Planning with R}}
 \description{
-The stplanr package provides functions to access and analyse data for transportation research, including origin-destination analysis, route allocation and modelling travel patterns.
+The stplanr package provides functions to access
+and analyse data for transportation research, including origin-destination analysis,
+route allocation and modelling travel patterns.
+}
+\section{Interesting functions}{
+
+\itemize{
+ \item \code{\link{gOverline}} - Aggregate overlaying route lines and data intelligently
+ \item \code{\link{calc_catchment}} - Create a 'catchment area' to show the areas serving a destination
+ \item \code{\link{route_cyclestreet}} - Finds the fastest routes for cyclists between two places.
+}
 }
 \author{
-\email{rob00x@gmail.com}
+Robin Lovelace \email{rob00x@gmail.com}
 }
 \seealso{
 \url{https://github.com/Robinlovelace/stplanr}
