### 2015-10-18 #10 @Robinlovelace
Thanks @sckott for this detailed review. I've pushed a load of changes that fix the following issues:
