### 2015-10-18 fdc02d1
Update read_table_builder examples to run
