### 2015-10-19 e021d47
Addition of calc_moving_catchment function and minor fixes to existing functions.
