+++ R/loadABS.R
@@ -26,7 +26,7 @@
 #' t1 <- read_table_builder(file.path(data_dir, 'SA1Population.csv'))
 #' t2 <- read_table_builder(file.path(data_dir, 'SA1Population.xlsx'),
 #'  filetype = 'xlsx', sheet = 1, removeTotal = TRUE)
-#' sa1pop <- read.csv(file.path(data_dir, 'SA1Population.csv'))
+#' sa1pop <- read.csv(file.path(data_dir, 'SA1Population.csv'), header=FALSE)
 #' t3 <- read_table_builder(sa1pop)
 read_table_builder <- function(dataset, filetype="csv",sheet=1,removeTotal=TRUE) {
   if (missing(dataset)) {
