+++ NAMESPACE
@@ -5,6 +5,7 @@ export(age_recat2)
 export(bbox_scale)
 export(calc_catchment)
 export(calc_catchment_sum)
+export(calc_moving_catchment)
 export(dd_logcub)
 export(dd_loglin)
 export(dd_logsqrt)
