### 2015-10-19 #10 @sckott
A bit more review:

* The `route_cyclestreet()` examples failed for me. The `ggmap::geocode` calls don't work:

    ```r
    route_cyclestreet(from = "Hereford", to = "Leeds", plan = "quietest")
    #> Information from URL : http://www.datasciencetoolkit.org/maps/api/geocode/json?address=Hereford&sensor=false
    #> Information from URL : http://www.datasciencetoolkit.org/maps/api/geocode/json?address=Leeds&sensor=false
    #> 1: In readLines(connect, warn = FALSE) :
    #>   cannot open: HTTP status was '0 (null)'
    #> 2: In ggmap::geocode(from) :   geocoding failed for "Hereford".
    #>   if accompanied by 500 Internal Server Error with using dsk, try google.
    #> 3: In readLines(connect, warn = FALSE) :
    #>   cannot open: HTTP status was '0 (null)'
    #> 4: In ggmap::geocode(to) :   geocoding failed for "Leeds".
    #>   if accompanied by 500 Internal Server Error with using dsk, try google.
    ```

    Is there something simple I'm missing. Do I need auth to use `ggmap::geocode`. Perhaps you could use `geonames` package?  E.g.,

    ```r
    geonames::GNpostalCodeSearch(postalcode="M3 4EE")
    #>   adminCode2 adminCode3              adminName3 adminCode1               lng countryCode postalCode adminName1        placeName
    #> 1             E08000003 Manchester District (B)        ENG -2.24808567051586          GB     M3 4EE    England City Centre Ward
    #>           lat
    #> 1 53.4772316872306
    ```

* Same for `route_graphhopper()`
