### 2015-10-19 #10 @sckott
> Please take a look at my new package description stplanr-package.R: seem OK?

Yep

> One problem is that ggmap::geocode seems unreliable so make look for a replacement. This comment is the biggy for me so any suggestions, let me know.

See my above comment, perhaps `geonames`

> Point to instructions in readme for installing rgdal and rgeos, they are often painful for folks
Fixed here for Ubuntu - suspect it's less of an issue for Windows users but I don't know how to fix this for Mac users

e.g, for macs https://github.com/ropensci/geojsonio#install

> Any advice re. how to prompt? Should it be a warning that it needs api keys on package load or only on the function calls?

Not on package load. You could also have a separate function to set keys, a wrapper around `Sys.setenv`

> gOverline crashing: Please see image below. What's your session_info?

after loading `stplanr`

    ```r
    Session info ---------------------------------
     setting  value
     version  R version 3.2.2 Patched (2015-08-14 r69078)
     system   x86_64, darwin13.4.0
     ui       RStudio (0.99.703)
     language (EN)
     collate  en_US.UTF-8
     tz       America/Los_Angeles
     date     2015-10-19

    Packages ---------------------------------------------
    package  * version date       source
     devtools * 1.9.1   2015-09-11 CRAN (R 3.2.0)
     digest     0.6.8   2014-12-31 CRAN (R 3.2.0)
     lattice    0.20-33 2015-07-14 CRAN (R 3.2.2)
     memoise    0.2.1   2014-04-22 CRAN (R 3.2.0)
     openxlsx   3.0.0   2015-07-03 CRAN (R 3.2.0)
     Rcpp       0.12.1  2015-09-10 CRAN (R 3.2.2)
     sp       * 1.2-0   2015-09-08 CRAN (R 3.2.2)
     stplanr  * 0.0.1.0 2015-10-16 Github (Robinlovelace/stplanr@0dac505)
    ```
