### 2015-10-20 e329884
Changes to gClip function to fix problems when the original layer has an "id" column in the data frame.
