+++ README.md
@@ -72,70 +72,8 @@ plot(travel_network, lwd = w)
 
 ![](README_files/figure-html/plot1-1.png) 
 
-The package can also allocate flows to the travel network, for example through
+The package can also allocate flows to the road network, for example through
 a link to the [CycleStreets.net API](https://www.cyclestreets.net/api/):
-
-
-```r
-example("line2route")
-```
-
-![](README_files/figure-html/plot2-1.png) ![](README_files/figure-html/plot2-2.png) 
-
-## Installation
-
-
-```r
-# you must have the devtools package (e.g. via install.packages("devtools"))
-devtools::install_github("robinlovelace/stplanr")
-library(stplanr)
-```
-
-stplanr depends on rgdal, which can be difficult to install, especially for Mac and Linux
-users.
-
-### Install binary version of rgdal
-
-This can be done from Ubuntu with
-
-```
-sudo apt-get install r-cran-rgdal
-```
-
-Also ggplot2, on which ggmap depends, may need to be installed as a binary, e.g.:
-
-```
-sudo apt-get install r-cran-ggplot2
-```
-
-### Set up rgdal manually
-
-The version of gdal needs to be newer than 1.11
-
-```r
-rgdal::getGDALVersionInfo()
-```
-
-```
-## [1] "GDAL 1.11.2, released 2015/02/10"
-```
-
-```r
-# Should return GDAL 1.11.2, released 2015/02/10 (or newer)
-```
-
-It is possible to use the following Personal Package Archive (PPA) to get the latest version of gdal
-
-
-```bash
-sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable && sudo apt-get update
-sudo apt-get install gdal-bin libgdal-dev
-```
-
-## Getting help
-
-We aim to make this package well-documented to make it easy to use.
-R's internal help functions will help here:
 
 
 ```r
@@ -154,10 +92,69 @@ tm_shape(osm_tiles) +
 ## unknown. Long-lat (WGS84) is assumed.
 ```
 
-![](README_files/figure-html/unnamed-chunk-7-1.png) 
+![](README_files/figure-html/cycle-trip-1.png) 
 
-The current list of available functions from stplanr is printed by the following
-command:
+We can replicate this call to CycleStreets.net multiple times
+using `line2route`.
+
+
+```r
+# Remove intra-zone flow
+intrazone <- travel_network$Area.of.residence == travel_network$Area.of.workplace
+travel_network <- travel_network[!intrazone,]
+routes <- line2route(travel_network)
+plot(routes)
+```
+
+![](README_files/figure-html/plot2-1.png) 
+
+For more examples, `example("line2route")`.
+
+`gOverline` is a function which takes a series of route-allocated lines,
+splits them into unique segmentes and aggregates
+the values of overlapping lines. This can represent where there will be
+most traffic on the transport system, as illustrated below.
+
+
+```r
+routes$All <- travel_network$All
+rnet <- gOverline(sldf = routes, attrib = "All", fun = sum)
+plot(rnet, lwd = rnet$All / mean(rnet$All))
+points(cents, col = "red", pch = 18)
+```
+
+![](README_files/figure-html/rnet-1.png) 
+
+## Installation
+
+
+```r
+# you must have the devtools package (e.g. via install.packages("devtools"))
+devtools::install_github("robinlovelace/stplanr")
+library(stplanr)
+```
+
+stplanr depends on rgdal, which can be difficult to installon Mac and Linux
+users. 
+
[…]
