### 2015-10-20 f245e12
Remove ggmap and RCurl dependencies
RgoogleMaps::getGeoCode replaces ggmap::geocode

Small updates to vignette
