+++ NAMESPACE
@@ -27,7 +27,7 @@ export(route_graphhopper)
 import(rgdal)
 import(rgeos)
 import(sp)
-importFrom(ggmap,geocode)
+importFrom(RgoogleMaps,getGeoCode)
 importFrom(jsonlite,fromJSON)
 importFrom(openxlsx,readWorkbook)
 importFrom(raster,crop)
