### 2015-10-20 225c02e
Improve robustness of vignette, rm ggmap dependency
