+++ vignettes/introducing-stplanr.Rmd
@@ -12,9 +12,7 @@ vignette: >
 ```{r, include=FALSE}
 library(stplanr)
 library(sp) # needed for geographical objects
-library(ggmap) # needed for plotting data
-library(rgdal) # required for spTransform
-library(RCurl)
+library(tmap) # needed for plotting data
 ```
 
 
@@ -125,6 +123,7 @@ an API key for the function to work):
 # test the connection
 x = try(download.file("http://www.google.com", "test.html"), silent = T) 
 offline <- grepl("Error", x)
+if(!offline) file.remove("test.html")
 
 data("routes_fast")
 data("routes_slow")
@@ -136,7 +135,7 @@ lines(routes_fast, col = "red")
 The above code sends requests to CycleStreets.net and saves the results.
 By changing the `plan` argument of `line2route`, we can download
 routes that are more suitable for people prioritising speed, quietness or
-a balance between speed and quietness. The plots of `l` and `rf` illustrate
+a balance between speed and quietness. The plots of `l` and `routes_fast` illustrate
 how the spatial classes that the flow data are converted to are easy to
 visualise. Using such data objects, generated by **stplanr**, in combination
 with additional packages for creating maps in R
@@ -197,21 +196,33 @@ are clearly allocated to the transport network
 (the package 'leaflet' was used to generate the basemap).
 
 
-```{r, echo=FALSE, warning=FALSE, message=FALSE, eval=FALSE}
+```{r, echo=FALSE, warning=FALSE, message=FALSE}
 if(offline){
   print("Figure not shown: requires internet connection")
 } else {
   n <- 30
-  lf <- fortify(l[n,])
-  rff <- fortify(rf[n,])
-  rqf <- fortify(rq[n,])
-  bb <- bbox(rf[n,])
-  bb <- make_bbox(lon = bb[1,], lat = bb[2,], f = 0.5)
-  ggmap(ggmap = get_map(location = bb)) +
-    geom_path(aes(x = long, y = lat), data = lf) + 
-    geom_path(aes(x = long, y = lat), data = rqf, colour = "red") + 
-    geom_path(aes(x = long, y = lat), data = rff, colour = "green") +
-    theme_nothing()
+  bb <- bbox(routes_fast[n,])
+  # lf <- fortify(l[n,])
+  # rff <- fortify(routes_fast[n,])
+  # rqf <- fortify(routes_slow[n,])
+  # the ggmap way
+  # bb <- make_bbox(lon = bb[1,], lat = bb[2,], f = 0.5)
+  # ggmap(ggmap = get_map(location = bb)) +
+  #   geom_path(aes(x = long, y = lat), data = lf) + 
+  #   geom_path(aes(x = long, y = lat), data = rqf, colour = "red") + 
+  #   geom_path(aes(x = long, y = lat), data = rff, colour = "green") +
+  #   theme_nothing()
+  
+  # the tmap way
+  otile <- read_osm(bbox(routes_fast))
+  tm_shape(otile) +
+    tm_raster() +
+    tm_shape(routes_fast) +
+    tm_lines() +
+    tm_shape(routes_slow) +
+    tm_lines(col = "darkgreen") +
+    tm_shape(l) +
+    tm_lines(col = "grey")
 }
 ```
 
