### 2015-10-20 4074a6e
Update vignette to include gOverline and leaflet
