### 2015-10-21 5d314bd
Update README to plot the flow on the basemap
Fixes this issue (that never was!) https://github.com/mtennekes/tmap/issues/35
