### 2015-10-21 29b8e93
Remove ggplot2 dependency
