### 2015-10-23 54abcc4
Add tests for routing functions
