### 2015-10-23 815afa2
Purge package of the evil camelCase
