+++ NAMESPACE
@@ -10,12 +10,12 @@ export(dd_logcub)
 export(dd_loglin)
 export(dd_logsqrt)
 export(disab_recat)
-export(gClip)
-export(gMapshape)
-export(gOnewaygeo)
-export(gOnewayid)
-export(gOverline)
-export(gSection)
+export(gclip)
+export(mapshape)
+export(onewaygeo)
+export(onewayid)
+export(overline)
+export(gsection)
 export(islines)
 export(line2df)
 export(line2route)

[…]
