### 2015-10-23 a6d8927
Fix authentication method
