### 2015-10-23 ed13a96
Add sp:: namespace to routes function
