### 2015-10-23 3e4d307
Update routes documentation for new api handling
