### 2015-10-23 6d9c47a
Small update to documenation
