### 2015-10-23 b3f8cb3
Add httr package
