+++ man/od2line.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/gFlow.R
+% Please edit documentation in R/od-funs.R
 \name{od2line}
 \alias{od2line}
 \title{Convert flow data to SpatialLinesDataFrame}
