+++ man/onewayid.Rd
@@ -1,21 +1,25 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/gOverline.R
-\name{gOnewaygeo}
-\alias{gOnewaygeo}
-\title{Aggregate flows so they become non-directional (by geometry - the slow way)}
+% Please edit documentation in R/overline.R
+\name{onewayid}
+\alias{onewayid}
+\title{Aggregate flows so they become non-directional (by id - the faster way)}
 \usage{
-gOnewaygeo(x, attrib)
+onewayid(x, attrib, id1 = names(x)[1], id2 = names(x)[2])
 }
 \arguments{
 \item{x}{A SpatialLinesDataFrame}
 
 \item{attrib}{A text string containing the name of the line's attribute to
 aggregate or a numeric vector of the columns to be aggregated}
+
+\item{id1}{A text string referring to the name of the variable containing the unique id of the origin}
+
+\item{id2}{A text string referring to the name of the variable containing the unique id of the destination}
 }
 \value{
-\code{gOnewaygeo} outputs a SpatialLinesDataFrame with single lines
+\code{onewayid} outputs a SpatialLinesDataFrame with single lines
 and user-selected attribute values that have been aggregated. Only lines
-with a distance (i.e. not intra-zone flows) are included
+with a distance (i.e. not intra-zone flows) are included.
 }
 \description{
 Flow data often contains movement in two directions: from point A to point B
@@ -35,11 +39,14 @@ of overlapping lines to zero.
 }
 \examples{
 data("flowlines")
+id1 <- names(flowlines)[1]
+id2 <- names(flowlines)[2]
 plot(flowlines)
-singlelines <- gOnewaygeo(flowlines, attrib = 3:14)
-plot(singlelines, lwd = 3, col = "red")
+singlelines <- onewayid(flowlines, attrib = 3:14, id1, id2)
 lines(singlelines) # check we've got the right lines
 sum(singlelines$All)
 nrow(singlelines)
+sl2 <- onewaygeo(flowlines, attrib = 3:14)
+# Demonstrate the results from onewayid and onewaygeo are identical
 }
 
