+++ man/line2df.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/gFlow.R
+% Please edit documentation in R/od-funs.R
 \name{line2df}
 \alias{line2df}
 \title{Convert straight SpatialLinesDataFrame to a data.frame with from and to coords}
