+++ man/gsection.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/gOverline.R
-\name{gSection}
-\alias{gSection}
+% Please edit documentation in R/overline.R
+\name{gsection}
+\alias{gsection}
 \title{Function to split overlapping SpatialLines into segments}
 \usage{
-gSection(sl)
+gsection(sl)
 }
 \arguments{
 \item{sl}{SpatialLinesDataFrame with overlapping Lines to split by
@@ -18,7 +18,7 @@ of aggregated lines.
 \examples{
 \dontrun{
 data(routes_fast)
-rsec <- gSection(routes_fast)
+rsec <- gsection(routes_fast)
 plot(routes_fast)
 lines(rsec, col = "red", lwd = 3)
 length(rsec)
