+++ man/lineLabels.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/gOverline.R
+% Please edit documentation in R/overline.R
 \name{lineLabels}
 \alias{lineLabels}
 \title{Label SpatialLinesDataFrame objects}
