+++ man/line2route.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/gFlow.R
+% Please edit documentation in R/od-funs.R
 \name{line2route}
 \alias{line2route}
 \title{Convert straight SpatialLinesDataFrame from flow data into routes}
