+++ man/gclip.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
 % Please edit documentation in R/geo-functions.R
-\name{gClip}
-\alias{gClip}
+\name{gclip}
+\alias{gclip}
 \title{Crops spatial object x to the bounding box of spatial object (or matrix) b}
 \usage{
-gClip(shp, bb)
+gclip(shp, bb)
 }
 \arguments{
 \item{shp}{The spatial object a to be cropped}
@@ -26,9 +26,9 @@ bb <- bbox(cents)
 cb <- rgeos::gBuffer(cents[8, ], width = 0.012, byid = TRUE)
 plot(cents)
 plot(cb, add = TRUE)
-clipped <- gClip(cents, cb)
+clipped <- gclip(cents, cb)
 row.names(clipped)
-clipped$avslope # gClip also returns the data attribute
+clipped$avslope # gclip also returns the data attribute
 points(clipped)
 points(cents[cb,], col = "red") # note difference
 }

[…]
