### 2015-10-23 #10 @Robinlovelace
I think all the outstanding issues except for the unit tests for all functions have been tackled. I added unit tests for more functions in a [commit today](https://github.com/Robinlovelace/stplanr/commit/54abcc45469a9a03102fe0d1073e2c10b3bdee34), but there are still functions that lack tests. The issues with ggmap have been resolved by use of `RgoogleMaps::getGeoCode()`. Addition I'm really proud of - R asks you for your api key interactively if its not set when you use the `route_` functions. Please give them a spin!
