+++ README.md
@@ -186,24 +186,26 @@ lsf.str("package:stplanr", all = TRUE)
 ##     retainAreaProportion = FALSE)  
 ## calc_moving_catchment : function (polygonlayer, targetlayer, calccols, distance = 500, projection = "worldalbers", 
 ##     retainAreaProportion = FALSE)  
+## cyclestreet_pat : function (force = FALSE)  
 ## dd_logcub : function (x, a, b1, b2, b3)  
 ## dd_loglin : function (x, a = 0.3, b1 = -0.2)  
 ## dd_logsqrt : function (x, a, b1, b2)  
 ## disab_recat : function (a)  
 ## gclip : function (shp, bb)  
-## mapshape : function (dsn, percent)  
-## onewaygeo : function (x, attrib)  
-## onewayid : function (x, attrib, id1 = names(x)[1], id2 = names(x)[2])  
-## overline : function (sldf, attrib, fun = sum, na.zero = FALSE)  
+## graphhopper_pat : function (force = FALSE)  
 ## gsection : function (sl)  
 ## islines : function (g1, g2)  
 ## line2df : function (l)  
 ## line2route : function (ldf, ...)  
 ## lineLabels : function (sldf, attrib)  
+## mapshape : function (dsn, percent)  
 ## od2line : function (flow, zones)  
+## onewaygeo : function (x, attrib)  
+## onewayid : function (x, attrib, id1 = names(x)[1], id2 = names(x)[2])  
+## overline : function (sldf, attrib, fun = sum, na.zero = FALSE)  
 ## read_table_builder : function (dataset, filetype = "csv", sheet = 1, removeTotal = TRUE)  
-## route_cyclestreet : function (from, to, plan = "fastest", silent = FALSE)  
-## route_graphhopper : function (from, to, vehicle = "bike")
+## route_cyclestreet : function (from, to, plan = "fastest", silent = TRUE, pat = cyclestreet_pat())  
+## route_graphhopper : function (from, to, vehicle = "bike", silent = TRUE, pat = graphhopper_pat())
 ```
 
 To get internal help on a specific function, use the standard way.
