### 2015-10-24 #10 @sckott
> I tested it but found it less effective

okay, thanks for checking

I'll give it a spin...
