### 2015-10-25 3b2b368
Improvements to calc_catchment functions, addition of data to extdata folder and correction of examples. Also added testthat files.
