+++ man/calc_moving_catchment.Rd
@@ -52,6 +52,11 @@ columns with summary variables.
 }
 \examples{
 \dontrun{
+data_dir <- system.file("extdata", package = "stplanr")
+unzip(file.path(data_dir, 'smallsa1.zip'))
+unzip(file.path(data_dir, 'testcycleway.zip'))
+sa1income <- readOGR(".","smallsa1")
+testcycleway <- readOGR(".","testcycleway")
 calc_moving_catchment(
    polygonlayer = sa1income,
    targetlayer = testcycleway,
