### 2015-10-26 #10 @sckott
@Robinlovelace Tested the routes functions that require keys. Works nicely now!
