### 2015-10-27 #10 @sckott
Sure, and nice work on the package.

Nothing major left I can think of, but installing with building vignette for me turns up an error `NOTE: rgdal::checkCRSArgs: no proj_defs.dat in PROJ.4 shared files` - which makes the install fail. Do you ever get this error?
