### 2015-10-27 #10 @Robinlovelace
Hi Scott, no I don't get that error message. Following advice from here, I get:

```
devtools::install_github("robinlovelace/stplanr", build_vignettes = T)
Downloading GitHub repo robinlovelace/stplanr@master
## lots of messages associated with install...
## ...
* DONE (stplanr)
Reloading installed stplanr
```

The only strange thing was this note but I don't think that's anything to do with stplanr:

```
Note: the specification for S3 class “AsIs” in package ‘jsonlite’ seems equivalent to one from package ‘RJSONIO’: not turning on duplicate class definitions for this class.
```

Regarding the ROpenSci process, what's the next stage? Do you recommend I submit it to CRAN?
