### 2015-10-27 #10 @sckott
I am getting a segfault though on the call to `overline()` https://github.com/Robinlovelace/stplanr/blame/master/vignettes/introducing-stplanr.Rmd#L202 Any thoughts?
