### 2015-10-27 #10 @Robinlovelace
I don't get any error on my Linux laptop. Just tried on my Windows work desktop: no issues either.
