### 2015-10-27 #10 @sckott
What version of GDAL/GEOS do you have?  I have gdal 1.11.1, geos 3.4.2, proj 4.9.2
