### 2015-10-27 #10 @Robinlovelace
I get
```
> rgdal::getGDALVersionInfo()
[1] "GDAL 1.11.2, released 2015/02/10"
```
on both Windows and Linux systems.
