### 2015-10-27 #10 @sckott
Thanks. Other people can't reproduce this bug, so we'll figure it out later if it persists, looks to be just osx, and just me so far :)

Just one more thing on a check I just ran, a few lines too long

```r
Rd file 'calc_catchment.Rd':
  \usage lines wider than 90 characters:
       projection = "+proj=aea +lat_1=90 +lat_2=-18.416667 +lat_0=0 +lon_0=10 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no ... [TRUNCATED]

Rd file 'calc_catchment_sum.Rd':
  \usage lines wider than 90 characters:
       projection = "+proj=aea +lat_1=90 +lat_2=-18.416667 +lat_0=0 +lon_0=10 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no ... [TRUNCATED]
```
