### 2015-10-27 #10 @sckott
After that, we're ready to bring it in to ropensci.
