### 2015-10-27 7a2cfdd
Split long char strings to avoid check note
