+++ man/calc_catchment.Rd
@@ -5,8 +5,10 @@
 \title{Calculate catchment area and associated summary statistics.}
 \usage{
 calc_catchment(polygonlayer, targetlayer, calccols, distance = 500,
-  projection = "+proj=aea +lat_1=90 +lat_2=-18.416667 +lat_0=0 +lon_0=10 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs",
-  retainAreaProportion = FALSE, dissolve = FALSE)
+  projection = paste0("+proj=aea +lat_1=90 +lat_2=-18.416667 ",
+  "+lat_0=0 +lon_0=10 +x_0=0 +y_0=0 +ellps=GRS80",
+  " +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"), retainAreaProportion = FALSE,
+  dissolve = FALSE)
 }
 \arguments{
 \item{polygonlayer}{A SpatialPolygonsDataFrame containing zones from which
