+++ man/calc_catchment_sum.Rd
@@ -5,7 +5,9 @@
 \title{Calculate summary statistics for catchment area.}
 \usage{
 calc_catchment_sum(polygonlayer, targetlayer, calccols, distance = 500,
-  projection = "+proj=aea +lat_1=90 +lat_2=-18.416667 +lat_0=0 +lon_0=10 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs",
+  projection = paste0("+proj=aea +lat_1=90 +lat_2=-18.416667",
+  " +lat_0=0 +lon_0=10 +x_0=0 +y_0=0",
+  " +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"),
   retainAreaProportion = FALSE)
 }
 \arguments{
