### 2015-10-27 #10 @Robinlovelace
Great. This fixes the 90 char limit note I think - suggest we're ready to go!
https://github.com/Robinlovelace/stplanr/commit/7a2cfddb2c20ad1c053de2ac556362f3b6fd4e18
