### 2015-10-27 9a2f725
Update CRAN comments
