### 2015-10-27 #10 @sckott
I made a team within ropensci org account, you should be able to transfer the repo to ropensci after you accept that invitation (i think)
