### 2015-10-27 #10 @Robinlovelace
Hi @sckott thanks for inviting me to @ropensci. I've accepted but cannot transfer stplanr. I think this is because don't have the right permissions, based on [this](https://help.github.com/articles/transferring-a-repository/): I cannot create new repos it seems. Maybe you need to give me more permissions? I had some reservations about transferring stplanr to ropensci but having checked out more about the organisation I think it's the right thing to do: encourages community participation and provides some visibility. The support you guys provide is fantastic and it's soooo good having someone review your code: awesome work on bringing open access peer review into R package development.
