### 2015-10-28 #10 @sckott
Thanks, we think reviewing is going pretty well :)

Did you go to Settings > Transfer Ownership (near the bottom) ? That didn't work?
